@@ -1,16 +1,12 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ShopApp.Business.Abstract
+namespace ShopApp.Business.Abstract;
+
+public interface ICartService
 {
-    public interface ICartService
-    {
-        void InitializeCart(string userId);
-        Cart GetCartByUserId(string userId);
-        void AddToCart(string userId, int productId, int quantity);
-        void DeleteFromCart(string userId, int productId);
-        void ClearCart(string cartId);
-    }
+    void InitializeCart(string userId);
+    Cart? GetCartByUserId(string userId);
+    void AddToCart(string userId, int productId, int quantity);
+    void DeleteFromCart(string userId, int productId);
+    void ClearCart(string cartId);
 }
