@@ -1,21 +1,17 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ShopApp.Business.Abstract
+namespace ShopApp.Business.Abstract;
+
+public interface IProductService : IValidator<Product>
 {
-   public interface IProductService: IValidator<Product>
-    {
-        Product GetById(int id);
-        Product GetProductDetails(int id);
-        List<Product> GetAll();
-        List<Product> GetProductsByCategory(string category,int page,int pageSize);
-        bool Create(Product entity);
-        void Update(Product entity);
-        void Delete(Product entity);
-        int GetCountByCategory(string category);
-        Product GetByIdWithCategories(int id);
-        void Update(Product entity, int[] categoryIds);
-    }
+    Product? GetById(int id);
+    Product? GetProductDetails(int id);
+    List<Product> GetAll();
+    List<Product> GetProductsByCategory(string category, int page, int pageSize);
+    bool Create(Product entity);
+    void Update(Product entity);
+    void Delete(Product entity);
+    int GetCountByCategory(string category);
+    Product? GetByIdWithCategories(int id);
+    void Update(Product entity, int[] categoryIds);
 }
