@@ -1,27 +1,25 @@
 ﻿using ShopApp.Business.Abstract;
 using ShopApp.DataAccess.Abstract;
 using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ShopApp.Business.Concrete
+namespace ShopApp.Business.Concrete;
+
+public class OrderManager : IOrderService
 {
-    public class OrderManager : IOrderService
-    {
-        private IOrderDal _orderDal;
-        public OrderManager(IOrderDal orderDal)
-        {
-            _orderDal = orderDal;
-        }
-        public void Create(Order entity)
-        {
-            _orderDal.Create(entity);
-        }
+    private readonly IOrderDal _orderDal;
 
-        public List<Order> GetOrders(string userId)
-        {
-            return _orderDal.GetOrders(userId);
-        }
+    public OrderManager(IOrderDal orderDal)
+    {
+        _orderDal = orderDal;
+    }
+
+    public void Create(Order entity)
+    {
+        _orderDal.Create(entity);
+    }
+
+    public List<Order> GetOrders(string? userId)
+    {
+        return _orderDal.GetOrders(userId);
     }
 }
