@@ -1,100 +1,95 @@
 ﻿using ShopApp.Business.Abstract;
 using ShopApp.DataAccess.Abstract;
-using ShopApp.DataAccess.Concrete.EfCore;
 using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-namespace ShopApp.Business.Concrete
+namespace ShopApp.Business.Concrete;
+
+public class ProductManager : IProductService
 {
-    public class ProductManager : IProductService
+    private readonly IProductDal _productDal;
+
+    public ProductManager(IProductDal productDal)
     {
-        private IProductDal _productDal;
-        public ProductManager(IProductDal productDal)
+        _productDal = productDal;
+    }
+
+    public string ErrorMessage { get; set; } = string.Empty;
+
+    public bool Create(Product entity)
+    {
+        if (Validate(entity))
         {
-            _productDal = productDal;
+            _productDal.Create(entity);
+            return true;
+        }
+        return false;
+    }
+
+    public void Delete(Product entity)
+    {
+        _productDal.Delete(entity);
+    }
+
+    public List<Product> GetAll()
+    {
+        return _productDal.GetAll();
+    }
+
+    public Product? GetById(int id)
+    {
+        return _productDal.GetById(id);
+    }
+
+    public Product? GetByIdWithCategories(int id)
+    {
+        return _productDal.GetByIdWithCategories(id);
+    }
+
+    public int GetCountByCategory(string category)
+    {
+        return _productDal.GetCountByCategory(category);
+    }
+
+    public Product? GetProductDetails(int id)
+    {
+        return _productDal.GetProductDetails(id);
+    }
+
+    public List<Product> GetProductsByCategory(string category, int page, int pageSize)
+    {
+        return _productDal.GetProductsByCategory(category, page, pageSize);
+    }
+
+    public void Update(Product entity)
+    {
+        _productDal.Update(entity);
+    }
+
+    public void Update(Product entity, int[] categoryIds)
+    {
+        _productDal.Update(entity, categoryIds);
+    }
+
+    public bool Validate(Product entity)
+    {
+        var isValid = true;
+
+        if (string.IsNullOrEmpty(entity.Name))
+        {
+            ErrorMessage += "ürün ismi girmelisiniz";
+            isValid = false;
+        }
+        if (string.IsNullOrEmpty(entity.ImageUrl))
+        {
+            ErrorMessage += "image url girmelisiniz";
+            isValid = false;
+        }
+        if (string.IsNullOrEmpty(entity.Description))
+        {
+            ErrorMessage += "açıklama girmelisiniz";
+            isValid = false;
         }
 
-        
-
-        public bool Create(Product entity)
-        {
-            if(Validate(entity))
-            {
-                _productDal.Create(entity);
-                return true;
-            }
-            return false;
-        }
-
-        public void Delete(Product entity)
-        {
-            _productDal.Delete(entity);
-        }
-
-        public List<Product> GetAll()
-        {
-            return _productDal.GetAll();
-        }
-
-        public Product GetById(int id)
-        {
-            return _productDal.GetById(id);
-        }
-
-        public Product GetByIdWithCategories(int id)
-        {
-            return _productDal.GetByIdWithCategories(id);
-        }
-
-        public int GetCountByCategory(string category)
-        {
-            return _productDal.GetCountByCategory(category);
-        }
-
-        public Product GetProductDetails(int id)
-        {
-            return _productDal.GetProductDetails(id); 
-        }
-
-        public List<Product> GetProductsByCategory(string category, int page, int pageSize)
-        {
-            return _productDal.GetProductsByCategory(category, page, pageSize);
-        }
-
-        public void Update(Product entity)
-        {
-            _productDal.Update(entity);
-        }
-
-        public void Update(Product entity, int[] categoryIds)
-        {
-            _productDal.Update(entity, categoryIds);
-        }
-        public string ErrorMessage { get; set; }
-        public bool Validate(Product entity)
-        {
-            var isValid = true;
-
-            if (string.IsNullOrEmpty(entity.Name))
-            {
-                ErrorMessage += "ürün ismi girmelisiniz";
-                isValid = false;
-            }
-            if (string.IsNullOrEmpty(entity.ImageUrl))
-            {
-                ErrorMessage += "image url girmelisiniz";
-                isValid = false;
-            }
-            if (string.IsNullOrEmpty(entity.Description))
-            {
-                ErrorMessage += "açıklama girmelisiniz";
-                isValid = false;
-            }
-
-            return isValid;
-        }
+        return isValid;
     }
 }
