@@ -1,14 +1,10 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ShopApp.DataAccess.Abstract
+namespace ShopApp.DataAccess.Abstract;
+
+public interface ICartDal : IRepository<Cart>
 {
-    public interface ICartDal : IRepository<Cart>
-    {
-        Cart GetByUserId(string userId);
-        void DeleteFromCart(int cartId, int productId);
-        void ClearCart(string cartId);
-    }
+    Cart? GetByUserId(string userId);
+    void DeleteFromCart(int cartId, int productId);
+    void ClearCart(string cartId);
 }
