@@ -1,12 +1,8 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace ShopApp.DataAccess.Abstract
+namespace ShopApp.DataAccess.Abstract;
+
+public interface IOrderDal : IRepository<Order>
 {
-    public interface IOrderDal : IRepository<Order>
-    {
-        List<Order> GetOrders(string userId);
-    }
+    List<Order> GetOrders(string? userId);
 }
