@@ -1,18 +1,12 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
 
-namespace ShopApp.DataAccess.Abstract
+namespace ShopApp.DataAccess.Abstract;
+
+public interface IProductDal : IRepository<Product>
 {
-   public interface IProductDal: IRepository<Product>
-    {
-        List<Product> GetProductsByCategory(string category, int page, int pageSize);
-        Product GetProductDetails(int id);
-        int GetCountByCategory(string category);
-        Product GetByIdWithCategories(int id);
-        void Update(Product entity, int[] categoryIds);
-    }
+    List<Product> GetProductsByCategory(string category, int page, int pageSize);
+    Product? GetProductDetails(int id);
+    int GetCountByCategory(string category);
+    Product? GetByIdWithCategories(int id);
+    void Update(Product entity, int[] categoryIds);
 }
