@@ -1,18 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
+﻿using System.Linq.Expressions;
 
-namespace ShopApp.DataAccess.Abstract
+namespace ShopApp.DataAccess.Abstract;
+
+public interface IRepository<T>
 {
-    public interface IRepository<T>
-    {
-        T GetById(int id);
-        T GetOne(Expression<Func<T, bool>> filter);
-        List<T> GetAll(Expression<Func<T, bool>> filter = null);
-        void Create(T entity);
-        void Update(T entity);
-        void Delete(T entity);
-    }
+    T? GetById(int id);
+    T? GetOne(Expression<Func<T, bool>> filter);
+    List<T> GetAll(Expression<Func<T, bool>>? filter = null);
+    void Create(T entity);
+    void Update(T entity);
+    void Delete(T entity);
 }
