@@ -1,34 +1,25 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using ShopApp.DataAccess.Abstract;
 using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text;
 
-namespace ShopApp.DataAccess.Concrete.EfCore
+namespace ShopApp.DataAccess.Concrete.EfCore;
+
+public class EfCoreCategoryDal : EfCoreGenericRepository<Category, ShopContext>, ICategoryDal
 {
-    public class EfCoreCategoryDal : EfCoreGenericRepository<Category, ShopContext>, ICategoryDal
+    public void DeleteFromCategory(int categoryId, int productId)
     {
-        public void DeleteFromCategory(int categoryId, int productId)
-        {
-            using (var context = new ShopContext())
-            {
-                var cmd = @"delete from ProductCategory where ProductId=@p0 And CategoryId=@p1";
-                context.Database.ExecuteSqlCommand(cmd, productId, categoryId);
-            }
-        }
-        public Category GetByIdWithProducts(int id)
-        {
-            using (var context = new ShopContext())
-            {
-                return context.Categories
-                        .Where(i => i.Id == id)
-                        .Include(i => i.ProductCategories)
-                        .ThenInclude(i => i.Product)
-                        .FirstOrDefault();
-            }
-        }
+        using var context = new ShopContext();
+        var cmd = $"DELETE FROM ProductCategory WHERE ProductId={productId} AND CategoryId={categoryId}";
+        context.Database.ExecuteSqlRaw(cmd);
+    }
+
+    public Category? GetByIdWithProducts(int id)
+    {
+        using var context = new ShopContext();
+        return context.Categories
+                .Where(i => i.Id == id)
+                .Include(i => i.ProductCategories)
+                .ThenInclude(i => i.Product)
+                .FirstOrDefault();
     }
 }
