@@ -1,61 +1,55 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 
-namespace ShopApp.DataAccess.Concrete.EfCore
+namespace ShopApp.DataAccess.Concrete.EfCore;
+
+public static class SeedDatabase
 {
-    public static class SeedDatabase
+    public static void Seed()
     {
-        public static void Seed()
+        using var context = new ShopContext();
+
+        if (!context.Database.GetPendingMigrations().Any())
         {
-
-            var context = new ShopContext();
-
-            if (context.Database.GetPendingMigrations().Count()==0)
+            if (!context.Categories.Any())
             {
-                if(context.Categories.Count()==0)
-                {
-                    context.Categories.AddRange(Categories);
-                }
-                if(context.Products.Count()==0)
-                {
-                    context.Products.AddRange(Products);
-                    context.AddRange(ProductCategory);
-                }
-                context.SaveChanges();
+                context.Categories.AddRange(Categories);
             }
+            if (!context.Products.Any())
+            {
+                context.Products.AddRange(Products);
+                context.AddRange(ProductCategories);
+            }
+            context.SaveChanges();
         }
-
-        private static Category[] Categories =
-        {
-            new Category() { Name = "Sebze"},
-            new Category() { Name = "Meyve"},
-            new Category() { Name = "Bakliyat"}
-
-        };
-
-        private static Product[] Products =
-        {
-            new Product(){Name = "Domates", Price=5, ImageUrl="1.jpg", Description= "<p>yüzde yüz yerli tire domat organik</p>"},
-            new Product(){Name = "Havuç", Price=3, ImageUrl="2.jpg", Description= "<p>yüzde yüz yerli tire havuç organik</p>"},
-            new Product(){Name = "Patates", Price=4, ImageUrl="3.jpg", Description= "<p>yüzde yüz yerli tire patates organik</p>"},
-            new Product(){Name = "Salatalık", Price=6, ImageUrl="4.jpg", Description= "<p>yüzde yüz yerli tire salatalık organik</p>"},
-            new Product(){Name = "Taze Fasulye", Price=10, ImageUrl="5.jpg", Description= "<p>yüzde yüz yerli tire fasulye organik</p>"},
-            new Product(){Name = "meyve", Price=3, ImageUrl="1.jpg", Description= "<p>yüzde yüz yerli tire domat organik</p>"},
-            new Product(){Name = "meyve", Price=3, ImageUrl="1.jpg", Description= "<p>yüzde yüz yerli tire domat organik</p>"}
-        };
-        private static ProductCategory[] ProductCategory =
-        {
-            new ProductCategory() { Product= Products[0],Category= Categories[0]},
-            new ProductCategory() { Product= Products[0],Category= Categories[2]},
-            new ProductCategory() { Product= Products[1],Category= Categories[0]},
-            new ProductCategory() { Product= Products[1],Category= Categories[1]},
-            new ProductCategory() { Product= Products[2],Category= Categories[0]},
-            new ProductCategory() { Product= Products[2],Category= Categories[2]},
-            new ProductCategory() { Product= Products[3],Category= Categories[1]}
-        };
     }
+
+    private static readonly Category[] Categories =
+    {
+        new() { Name = "Sebze" },
+        new() { Name = "Meyve" },
+        new() { Name = "Bakliyat" }
+    };
+
+    private static readonly Product[] Products =
+    {
+        new() { Name = "Domates", Price = 5, ImageUrl = "1.jpg", Description = "<p>yüzde yüz yerli tire domat organik</p>" },
+        new() { Name = "Havuç", Price = 3, ImageUrl = "2.jpg", Description = "<p>yüzde yüz yerli tire havuç organik</p>" },
+        new() { Name = "Patates", Price = 4, ImageUrl = "3.jpg", Description = "<p>yüzde yüz yerli tire patates organik</p>" },
+        new() { Name = "Salatalık", Price = 6, ImageUrl = "4.jpg", Description = "<p>yüzde yüz yerli tire salatalık organik</p>" },
+        new() { Name = "Taze Fasulye", Price = 10, ImageUrl = "5.jpg", Description = "<p>yüzde yüz yerli tire fasulye organik</p>" },
+        new() { Name = "meyve", Price = 3, ImageUrl = "1.jpg", Description = "<p>yüzde yüz yerli tire domat organik</p>" },
+        new() { Name = "meyve", Price = 3, ImageUrl = "1.jpg", Description = "<p>yüzde yüz yerli tire domat organik</p>" }
+    };
+
+    private static readonly ProductCategory[] ProductCategories =
+    {
+        new() { Product = Products[0], Category = Categories[0] },
+        new() { Product = Products[0], Category = Categories[2] },
+        new() { Product = Products[1], Category = Categories[0] },
+        new() { Product = Products[1], Category = Categories[1] },
+        new() { Product = Products[2], Category = Categories[0] },
+        new() { Product = Products[2], Category = Categories[2] },
+        new() { Product = Products[3], Category = Categories[1] }
+    };
 }
