@@ -1,269 +1,274 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI.Services;
+﻿using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using ShopApp.Business.Abstract;
+using ShopApp.WebUI.EmailServices;
 using ShopApp.WebUI.Extensions;
 using ShopApp.WebUI.Identity;
 using ShopApp.WebUI.Models;
 
-namespace ShopApp.WebUI.Controllers
+namespace ShopApp.WebUI.Controllers;
+
+[AutoValidateAntiforgeryToken]
+public class AccountController : Controller
 {
-    [AutoValidateAntiforgeryToken]
-    public class AccountController : Controller
+    private readonly UserManager<ApplicationUser> _userManager;
+    private readonly SignInManager<ApplicationUser> _signInManager;
+    private readonly IEmailSender _emailSender;
+    private readonly ICartService _cartService;
+
+    public AccountController(
+        ICartService cartService, 
+        UserManager<ApplicationUser> userManager, 
+        SignInManager<ApplicationUser> signInManager, 
+        IEmailSender emailSender)
     {
-        private UserManager<ApplicationUser> _userManager;
-        private SignInManager<ApplicationUser> _signInManager;
-        private IEmailSender _emailSender;
-        private ICartService _cartService;
-        public AccountController(ICartService cartService, UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager, IEmailSender emailSender)
+        _cartService = cartService;
+        _userManager = userManager;
+        _signInManager = signInManager;
+        _emailSender = emailSender;
+    }
+
+    public IActionResult Register()
+    {
+        return View(new RegisterModel());
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> Register(RegisterModel model)
+    {
+        if (!ModelState.IsValid)
         {
-            _cartService = cartService;
-            _userManager = userManager;
-            _signInManager = signInManager;
-            _emailSender = emailSender;
-        }
-        public IActionResult Register()
-        {
-            return View(new RegisterModel());
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Register(RegisterModel model)
-        {
-            if (!ModelState.IsValid)
-            {
-                return View(model);
-            }
-
-            var user = new ApplicationUser
-            {
-                UserName = model.Username,
-                Email = model.Email,
-                FullName = model.FullName
-            };
-
-            var result = await _userManager.CreateAsync(user, model.Password);
-
-            if (result.Succeeded)
-            {
-                // generate token
-                var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
-                var callbackUrl = Url.Action("ConfirmEmail", "Account", new
-                {
-                    userId = user.Id,
-                    token = code
-                });
-
-                // send email
-                await _emailSender.SendEmailAsync(model.Email, "Hesabınızı Onaylayınız.", $"Lütfen email hesabınızı onaylamak için linke <a href='http://localhost:55041{callbackUrl}'>tıklayınız.</a>");
-
-                TempData.Put("message", new ResultMessage()
-                {
-                    Title = "Hesap Onayı",
-                    Message = "Eposta adrenize gelen link ile hesabınızı onaylayınız",
-                    Css = "warning"
-                });
-
-                return RedirectToAction("Login", "Account");
-            }
-
-
-
-            ModelState.AddModelError("", "Bilinmeyen hata oluştu lütfen tekrar deneyiniz.");          
             return View(model);
         }
 
-
-        public IActionResult Login(string ReturnUrl = null)
+        var user = new ApplicationUser
         {
-            return View(new LoginModel()
-            {
-                ReturnUrl = ReturnUrl
-            });
-        }
+            UserName = model.Username,
+            Email = model.Email,
+            FullName = model.FullName ?? string.Empty
+        };
 
-        [HttpPost]
-        public async Task<IActionResult> Login(LoginModel model)
+        var result = await _userManager.CreateAsync(user, model.Password ?? string.Empty);
+
+        if (result.Succeeded)
         {
-            if (!ModelState.IsValid)
+            // generate token
+            var code = await _userManager.GenerateEmailConfirmationTokenAsync(user);
+            var callbackUrl = Url.Action("ConfirmEmail", "Account", new
             {
-                return View(model);
-            }
-            var user = await _userManager.FindByEmailAsync(model.Email);
-
-            if (user == null)
-            {
-                ModelState.AddModelError("", "Bu email ile daha önce hesap oluşturulmamış.");
-                return View(model);
-            }
-
-            if (!await _userManager.IsEmailConfirmedAsync(user))
-            {
-                ModelState.AddModelError("", "Lütfen hesabınızı email ile onaylayınız.");
-                return View(model);
-            }
-
-
-            var result = await _signInManager.PasswordSignInAsync(user, model.Password, true, false);
-
-            if (result.Succeeded)
-            {
-                return Redirect(model.ReturnUrl ?? "~/");
-            }
-
-            ModelState.AddModelError("", "Email veya parola yanlış");
-            return View(model);
-        }
-
-        public async Task<IActionResult> Logout()
-        {
-            await _signInManager.SignOutAsync();
-
-            TempData.Put("message", new ResultMessage()
-            {
-                Title = "Oturum Kapatıldı.",
-                Message = "Hesabınız güvenli bir şekilde sonlandırıldı.",
-                Css = "warning"
-            });
-            return Redirect("~/");
-        }
-
-        public async Task<IActionResult> ConfirmEmail(string userId, string token)
-        {
-            if (userId == null || token == null)
-            {
-                TempData.Put("message", new ResultMessage()
-                {
-                    Title = "Hesap Onayı",
-                    Message = "Hesap onayı için bilgileriniz yanlış",
-                    Css = "danger"
-                });
-
-                return Redirect("~/");
-            }
-
-            var user = await _userManager.FindByIdAsync(userId);
-            if (user != null)
-            {
-                var result = await _userManager.ConfirmEmailAsync(user, token);
-                if (result.Succeeded)
-                {
-                    _cartService.InitializeCart(user.Id);
-                    TempData.Put("message", new ResultMessage()
-                    {
-                        Title = "Hesap Onayı",
-                        Message = "Hesabınız başarıyla onaylanmıştır.",
-                        Css = "success"
-                    });
-
-                    return RedirectToAction("Login");
-                }
-            }
-
-            TempData.Put("message", new ResultMessage()
-            {
-                Title = "Hesap Onayı",
-                Message = "Hesabınız onaylanamadı.",
-                Css = "danger"
-            });
-            return View();
-        }
-        public IActionResult ForgotPassword()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> ForgotPassword(string Email)
-        {
-            if (string.IsNullOrEmpty(Email))
-            {
-                TempData.Put("message", new ResultMessage()
-                {
-                    Title = "Forgot Password",
-                    Message = "Bilgileriniz Hatalı",
-                    Css = "danger"
-                });
-
-                return View();
-            }
-
-            var user = await _userManager.FindByEmailAsync(Email);
-
-            if (user == null)
-            {
-                TempData.Put("message", new ResultMessage()
-                {
-                    Title = "Forgot Password",
-                    Message = "Eposta adresi ile bir kullanıcı bulunamadı",
-                    Css = "danger"
-                });
-
-                return View();
-            }
-
-            var code = await _userManager.GeneratePasswordResetTokenAsync(user);
-
-            var callbackUrl = Url.Action("ResetPassword", "Account", new
-            {
+                userId = user.Id,
                 token = code
             });
 
             // send email
-            await _emailSender.SendEmailAsync(Email, "Reset Password", $"Parolanızı yenilemek için linke <a href='http://localhost:55041{callbackUrl}'>tıklayınız.</a>");
+            var baseUrl = $"{Request.Scheme}://{Request.Host}";
+            await _emailSender.SendEmailAsync(
+                model.Email ?? string.Empty, 
+                "Hesabınızı Onaylayınız.", 
+                $"Lütfen email hesabınızı onaylamak için linke <a href='{baseUrl}{callbackUrl}'>tıklayınız.</a>");
 
             TempData.Put("message", new ResultMessage()
             {
-                Title = "Forgot Password",
-                Message = "Parola yenilemek için hesabınıza mail gönderildi.",
+                Title = "Hesap Onayı",
+                Message = "Eposta adrenize gelen link ile hesabınızı onaylayınız",
                 Css = "warning"
             });
 
             return RedirectToAction("Login", "Account");
         }
 
+        ModelState.AddModelError("", "Bilinmeyen hata oluştu lütfen tekrar deneyiniz.");
+        return View(model);
+    }
 
-        public IActionResult ResetPassword(string token)
+    public IActionResult Login(string? ReturnUrl = null)
+    {
+        return View(new LoginModel()
         {
-            if (token == null)
-            {
-                return RedirectToAction("Home", "Index");
-            }
-            var model = new ResetPasswordModel { Token = token };
+            ReturnUrl = ReturnUrl
+        });
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> Login(LoginModel model)
+    {
+        if (!ModelState.IsValid)
+        {
             return View(model);
         }
 
-        [HttpPost]
-        public async Task<IActionResult> ResetPassword(ResetPasswordModel model)
+        var user = await _userManager.FindByEmailAsync(model.Email ?? string.Empty);
+
+        if (user == null)
         {
-            if (!ModelState.IsValid)
+            ModelState.AddModelError("", "Bu email ile daha önce hesap oluşturulmamış.");
+            return View(model);
+        }
+
+        if (!await _userManager.IsEmailConfirmedAsync(user))
+        {
+            ModelState.AddModelError("", "Lütfen hesabınızı email ile onaylayınız.");
+            return View(model);
+        }
+
+        var result = await _signInManager.PasswordSignInAsync(user, model.Password ?? string.Empty, true, false);
+
+        if (result.Succeeded)
+        {
+            return Redirect(model.ReturnUrl ?? "~/");
+        }
+
+        ModelState.AddModelError("", "Email veya parola yanlış");
+        return View(model);
+    }
+
+    public async Task<IActionResult> Logout()
+    {
+        await _signInManager.SignOutAsync();
+
+        TempData.Put("message", new ResultMessage()
+        {
+            Title = "Oturum Kapatıldı.",
+            Message = "Hesabınız güvenli bir şekilde sonlandırıldı.",
+            Css = "warning"
+        });
+        return Redirect("~/");
+    }
+
+    public async Task<IActionResult> ConfirmEmail(string? userId, string? token)
+    {
+        if (userId == null || token == null)
+        {
+            TempData.Put("message", new ResultMessage()
             {
-                return View(model);
-            }
+                Title = "Hesap Onayı",
+                Message = "Hesap onayı için bilgileriniz yanlış",
+                Css = "danger"
+            });
 
-            var user = await _userManager.FindByEmailAsync(model.Email);
-            if (user == null)
-            {
-                return RedirectToAction("Home", "Index");
-            }
+            return Redirect("~/");
+        }
 
-            var result = await _userManager.ResetPasswordAsync(user, model.Token, model.Password);
-
+        var user = await _userManager.FindByIdAsync(userId);
+        if (user != null)
+        {
+            var result = await _userManager.ConfirmEmailAsync(user, token);
             if (result.Succeeded)
             {
-                return RedirectToAction("Login", "Account");
+                _cartService.InitializeCart(user.Id);
+                TempData.Put("message", new ResultMessage()
+                {
+                    Title = "Hesap Onayı",
+                    Message = "Hesabınız başarıyla onaylanmıştır.",
+                    Css = "success"
+                });
+
+                return RedirectToAction("Login");
             }
+        }
+
+        TempData.Put("message", new ResultMessage()
+        {
+            Title = "Hesap Onayı",
+            Message = "Hesabınız onaylanamadı.",
+            Css = "danger"
+        });
+        return View();
+    }
+
+    public IActionResult ForgotPassword()
+    {
+        return View();
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> ForgotPassword(string Email)
+    {
+        if (string.IsNullOrEmpty(Email))
+        {
+            TempData.Put("message", new ResultMessage()
+            {
+                Title = "Forgot Password",
+                Message = "Bilgileriniz Hatalı",
+                Css = "danger"
+            });
+
+            return View();
+        }
+
+        var user = await _userManager.FindByEmailAsync(Email);
+
+        if (user == null)
+        {
+            TempData.Put("message", new ResultMessage()
+            {
+                Title = "Forgot Password",
+                Message = "Eposta adresi ile bir kullanıcı bulunamadı",
+                Css = "danger"
+            });
+
+            return View();
+        }
+
+        var code = await _userManager.GeneratePasswordResetTokenAsync(user);
+
+        var callbackUrl = Url.Action("ResetPassword", "Account", new
+        {
+            token = code
+        });
+
+        // send email
+        var baseUrl = $"{Request.Scheme}://{Request.Host}";
+        await _emailSender.SendEmailAsync(
+            Email, 
+            "Reset Password", 
+            $"Parolanızı yenilemek için linke <a href='{baseUrl}{callbackUrl}'>tıklayınız.</a>");
+
+        TempData.Put("message", new ResultMessage()
+        {
+            Title = "Forgot Password",
+            Message = "Parola yenilemek için hesabınıza mail gönderildi.",
+            Css = "warning"
+        });
+
+        return RedirectToAction("Login", "Account");
+    }
+
+    public IActionResult ResetPassword(string? token)
+    {
+        if (token == null)
+        {
+            return RedirectToAction("Index", "Home");
+        }
+        var model = new ResetPasswordModel { Token = token };
+        return View(model);
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> ResetPassword(ResetPasswordModel model)
+    {
+        if (!ModelState.IsValid)
+        {
             return View(model);
         }
 
-
-        public IActionResult Accessdenied()
+        var user = await _userManager.FindByEmailAsync(model.Email ?? string.Empty);
+        if (user == null)
         {
-            return View();
+            return RedirectToAction("Index", "Home");
         }
+
+        var result = await _userManager.ResetPasswordAsync(user, model.Token ?? string.Empty, model.Password ?? string.Empty);
+
+        if (result.Succeeded)
+        {
+            return RedirectToAction("Login", "Account");
+        }
+        return View(model);
+    }
+
+    public IActionResult Accessdenied()
+    {
+        return View();
     }
 }
