@@ -1,219 +1,217 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using ShopApp.Business.Abstract;
 using ShopApp.Entities;
 using ShopApp.WebUI.Models;
 
-namespace ShopApp.WebUI.Controllers
+namespace ShopApp.WebUI.Controllers;
+
+[Authorize(Roles = "admin")]
+public class AdminController : Controller
 {
-    [Authorize(Roles = "admin")]
-    public class AdminController : Controller
+    private readonly IProductService _productService;
+    private readonly ICategoryService _categoryService;
+
+    public AdminController(IProductService productService, ICategoryService categoryService)
     {
-        private IProductService _productService;
-        private ICategoryService _categoryService;
-        public AdminController(IProductService productService, ICategoryService categoryService)
+        _productService = productService;
+        _categoryService = categoryService;
+    }
+
+    public IActionResult ProductList()
+    {
+        return View(new ProductListModel()
         {
-            _productService = productService;
-            _categoryService = categoryService;
-        }
-        public IActionResult ProductList()
+            Products = _productService.GetAll()
+        });
+    }
+
+    [HttpGet]
+    public IActionResult CreateProduct()
+    {
+        return View(new ProductModel());
+    }
+
+    [HttpPost]
+    public IActionResult CreateProduct(ProductModel model)
+    {
+        if (ModelState.IsValid)
         {
-            return View(new ProductListModel()
+            var entity = new Product()
             {
-                Products = _productService.GetAll()
-            });
-        }
-
-        [HttpGet]
-        public IActionResult CreateProduct()
-        {
-            return View(new ProductModel());
-        }
-        //VALİDATION       **********************************************************************
-        [HttpPost]
-        public IActionResult CreateProduct(ProductModel model)
-        {
-            if (ModelState.IsValid)
-            {
-                var entity = new Product()
-                {
-                    Name = model.Name,
-                    Price = model.Price,
-                    Description = model.Description,
-                    ImageUrl = model.ImageUrl
-                };
-
-                if(_productService.Create(entity))
-                {
-                    return RedirectToAction("ProductList");
-                }
-                ViewBag.ErrorMessage = _productService.ErrorMessage;
-                return View(model);
-            }
-
-            return View(model);
-
-        }
-
-        public IActionResult EditProduct(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-            var entity = _productService.GetByIdWithCategories((int)id);
-
-            if (entity == null)
-            {
-                return NotFound();
-            }
-
-            var model = new ProductModel()
-            {
-                Id = entity.Id,
-                Name = entity.Name,
-                Price = entity.Price,
-                Description = entity.Description,
-                ImageUrl = entity.ImageUrl,
-                SelectedCategories = entity.ProductCategories.Select(i => i.Category).ToList()
+                Name = model.Name,
+                Price = model.Price,
+                Description = model.Description,
+                ImageUrl = model.ImageUrl
             };
 
-            ViewBag.Categories = _categoryService.GetAll();
-
-            return View(model);
-        }
-        // editlendi                                                                                      editlendi
-        [HttpPost]
-        public async Task<IActionResult> EditProduct(ProductModel model, int[] categoryIds, IFormFile file)
-        {
-            if (ModelState.IsValid)
+            if (_productService.Create(entity))
             {
-                var entity = _productService.GetById(model.Id);
-
-                if (entity == null)
-                {
-                    return NotFound();
-                }
-
-                entity.Name = model.Name;
-                entity.Description = model.Description;
-                entity.Price = model.Price;
-
-                if (file != null)
-                {
-                    entity.ImageUrl = file.FileName;
-
-                    var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\img", file.FileName);
-                    using (var stream = new FileStream(path, FileMode.Create))
-                    {
-                        await file.CopyToAsync(stream);
-                    }
-                }
-
-                _productService.Update(entity, categoryIds);
-
                 return RedirectToAction("ProductList");
             }
-
-            ViewBag.Categories = _categoryService.GetAll();
-
+            ViewBag.ErrorMessage = _productService.ErrorMessage;
             return View(model);
         }
 
-        [HttpPost]
-        public IActionResult DeleteProduct(int productId)
+        return View(model);
+    }
+
+    public IActionResult EditProduct(int? id)
+    {
+        if (id == null)
         {
-            var entity = _productService.GetById(productId);
+            return NotFound();
+        }
+        var entity = _productService.GetByIdWithCategories((int)id);
 
-            if (entity != null)
-            {
-                _productService.Delete(entity);
-            }
-
-            return RedirectToAction("ProductList");
+        if (entity == null)
+        {
+            return NotFound();
         }
 
-        public IActionResult CategoryList()
+        var model = new ProductModel()
         {
-            return View(new CategoryListModel()
-            {
-                Categories = _categoryService.GetAll()
-            });
-        }
-        [HttpGet]
-        public IActionResult CreateCategory()
-        {
-            return View(new CategoryModel());
-        }
-        // KATEGORİ VALİDATİON 777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777777
-        [HttpPost]
-        public IActionResult CreateCategory(CategoryModel model)
-        {
-            if (ModelState.IsValid)
-            {
-                var entity = new Category()
-                {
-                    Name = model.Name
-                };
-                _categoryService.Create(entity);
+            Id = entity.Id,
+            Name = entity.Name,
+            Price = entity.Price,
+            Description = entity.Description ?? string.Empty,
+            ImageUrl = entity.ImageUrl ?? string.Empty,
+            SelectedCategories = entity.ProductCategories.Select(i => i.Category).ToList()
+        };
 
-                return RedirectToAction("CategoryList");
-            }
-            return View(model);
-            
-        }
-        [HttpGet]
-        public IActionResult EditCategory(int id)
-        {
-            var entity = _categoryService.GetByIdWithProducts(id);
+        ViewBag.Categories = _categoryService.GetAll();
 
-            return View(new CategoryModel()
-            {
-                Id = entity.Id,
-                Name = entity.Name,
-                Products = entity.ProductCategories.Select(p => p.Product).ToList()
-            });
-        }
-       
-        [HttpPost]
-        public IActionResult EditCategory(CategoryModel model)
+        return View(model);
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> EditProduct(ProductModel model, int[] categoryIds, IFormFile? file)
+    {
+        if (ModelState.IsValid)
         {
-            var entity = _categoryService.GetById(model.Id);
+            var entity = _productService.GetById(model.Id);
+
             if (entity == null)
             {
                 return NotFound();
             }
 
             entity.Name = model.Name;
-            _categoryService.Update(entity);
+            entity.Description = model.Description;
+            entity.Price = model.Price;
 
-            return RedirectToAction("CategoryList");
-        }
-
-        [HttpPost]
-        public IActionResult DeleteCategory(int categoryId)
-        {
-            var entity = _categoryService.GetById(categoryId);
-
-            if (entity != null)
+            if (file != null)
             {
-                _categoryService.Delete(entity);
+                entity.ImageUrl = file.FileName;
+
+                var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "img", file.FileName);
+                using var stream = new FileStream(path, FileMode.Create);
+                await file.CopyToAsync(stream);
             }
 
-            return RedirectToAction("CategoryList");
+            _productService.Update(entity, categoryIds);
+
+            return RedirectToAction("ProductList");
         }
 
-        [HttpPost]
-        public IActionResult DeleteFromCategory(int categoryId, int productId)
+        ViewBag.Categories = _categoryService.GetAll();
+
+        return View(model);
+    }
+
+    [HttpPost]
+    public IActionResult DeleteProduct(int productId)
+    {
+        var entity = _productService.GetById(productId);
+
+        if (entity != null)
         {
-            _categoryService.DeleteFromCategory(categoryId, productId);
-            return Redirect("/admin/editcategory/" + categoryId);
+            _productService.Delete(entity);
         }
+
+        return RedirectToAction("ProductList");
+    }
+
+    public IActionResult CategoryList()
+    {
+        return View(new CategoryListModel()
+        {
+            Categories = _categoryService.GetAll()
+        });
+    }
+
+    [HttpGet]
+    public IActionResult CreateCategory()
+    {
+        return View(new CategoryModel());
+    }
+
+    [HttpPost]
+    public IActionResult CreateCategory(CategoryModel model)
+    {
+        if (ModelState.IsValid)
+        {
+            var entity = new Category()
+            {
+                Name = model.Name
+            };
+            _categoryService.Create(entity);
+
+            return RedirectToAction("CategoryList");
+        }
+        return View(model);
+    }
+
+    [HttpGet]
+    public IActionResult EditCategory(int id)
+    {
+        var entity = _categoryService.GetByIdWithProducts(id);
+
+        if (entity == null)
+        {
+            return NotFound();
+        }
+
+        return View(new CategoryModel()
+        {
+            Id = entity.Id,
+            Name = entity.Name,
+            Products = entity.ProductCategories.Select(p => p.Product).ToList()
+        });
+    }
+
+    [HttpPost]
+    public IActionResult EditCategory(CategoryModel model)
+    {
+        var entity = _categoryService.GetById(model.Id);
+        if (entity == null)
+        {
+            return NotFound();
+        }
+
+        entity.Name = model.Name;
+        _categoryService.Update(entity);
+
+        return RedirectToAction("CategoryList");
+    }
+
+    [HttpPost]
+    public IActionResult DeleteCategory(int categoryId)
+    {
+        var entity = _categoryService.GetById(categoryId);
+
+        if (entity != null)
+        {
+            _categoryService.Delete(entity);
+        }
+
+        return RedirectToAction("CategoryList");
+    }
+
+    [HttpPost]
+    public IActionResult DeleteFromCategory(int categoryId, int productId)
+    {
+        _categoryService.DeleteFromCategory(categoryId, productId);
+        return Redirect("/admin/editcategory/" + categoryId);
     }
 }
