@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using IyzipayCore;
-using IyzipayCore.Model;
-using IyzipayCore.Request;
+﻿using Iyzipay;
+using Iyzipay.Model;
+using Iyzipay.Request;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -13,25 +9,84 @@ using ShopApp.Entities;
 using ShopApp.WebUI.Identity;
 using ShopApp.WebUI.Models;
 
-namespace ShopApp.WebUI.Controllers
-{
-    [Authorize]
-    public class CartController : Controller
-    {
-        private ICartService _cartService;
-        private IOrderService _orderService;
-        private UserManager<ApplicationUser> _userManager;
+namespace ShopApp.WebUI.Controllers;
 
-        public CartController(IOrderService orderService, ICartService cartService, UserManager<ApplicationUser> userManager)
+[Authorize]
+public class CartController : Controller
+{
+    private readonly ICartService _cartService;
+    private readonly IOrderService _orderService;
+    private readonly UserManager<ApplicationUser> _userManager;
+    private readonly IConfiguration _configuration;
+
+    public CartController(
+        IOrderService orderService, 
+        ICartService cartService, 
+        UserManager<ApplicationUser> userManager,
+        IConfiguration configuration)
+    {
+        _orderService = orderService;
+        _cartService = cartService;
+        _userManager = userManager;
+        _configuration = configuration;
+    }
+
+    public IActionResult Index()
+    {
+        var userId = _userManager.GetUserId(User);
+        if (userId == null) return RedirectToAction("Login", "Account");
+
+        var cart = _cartService.GetCartByUserId(userId);
+        if (cart == null) return View(new CartModel());
+
+        return View(new CartModel()
         {
-            _orderService = orderService;
-            _cartService = cartService;
-            _userManager = userManager;
+            CartId = cart.Id,
+            CartItems = cart.CartItems.Select(i => new CartItemModel()
+            {
+                CartItemId = i.Id,
+                ProductId = i.Product.Id,
+                Name = i.Product.Name,
+                Price = i.Product.Price ?? 0,
+                ImageUrl = i.Product.ImageUrl,
+                Quantity = i.Quantity
+            }).ToList()
+        });
+    }
+
+    [HttpPost]
+    public IActionResult AddToCart(int productId, int quantity)
+    {
+        var userId = _userManager.GetUserId(User);
+        if (userId != null)
+        {
+            _cartService.AddToCart(userId, productId, quantity);
         }
-        public IActionResult Index()
+        return RedirectToAction("Index");
+    }
+
+    [HttpPost]
+    public IActionResult DeleteFromCart(int productId)
+    {
+        var userId = _userManager.GetUserId(User);
+        if (userId != null)
         {
-            var cart = _cartService.GetCartByUserId(_userManager.GetUserId(User));
-            return View(new CartModel()
+            _cartService.DeleteFromCart(userId, productId);
+        }
+        return RedirectToAction("Index");
+    }
+
+    public IActionResult Checkout()
+    {
+        var userId = _userManager.GetUserId(User);
+        if (userId == null) return RedirectToAction("Login", "Account");
+
+        var cart = _cartService.GetCartByUserId(userId);
+        if (cart == null) return RedirectToAction("Index");
+
+        var orderModel = new OrderModel
+        {
+            CartModel = new CartModel()
             {
                 CartId = cart.Id,
                 CartItems = cart.CartItems.Select(i => new CartItemModel()
@@ -39,34 +94,28 @@ namespace ShopApp.WebUI.Controllers
                     CartItemId = i.Id,
                     ProductId = i.Product.Id,
                     Name = i.Product.Name,
-                    Price = (decimal)i.Product.Price,
+                    Price = i.Product.Price ?? 0,
                     ImageUrl = i.Product.ImageUrl,
                     Quantity = i.Quantity
                 }).ToList()
-            });
-        }
+            }
+        };
 
-        [HttpPost]
-        public IActionResult AddToCart(int productId, int quantity)
+        return View(orderModel);
+    }
+
+    [HttpPost]
+    public IActionResult Checkout(OrderModel model)
+    {
+        if (ModelState.IsValid)
         {
-            _cartService.AddToCart(_userManager.GetUserId(User), productId, quantity);
-            return RedirectToAction("Index");
-        }
+            var userId = _userManager.GetUserId(User);
+            if (userId == null) return RedirectToAction("Login", "Account");
 
-        [HttpPost]
-        public IActionResult DeleteFromCart(int productId)
-        {
-            _cartService.DeleteFromCart(_userManager.GetUserId(User), productId);
-            return RedirectToAction("Index");
-        }
+            var cart = _cartService.GetCartByUserId(userId);
+            if (cart == null) return RedirectToAction("Index");
 
-        public IActionResult Checkout()
-        {
-            var cart = _cartService.GetCartByUserId(_userManager.GetUserId(User));
-
-            var orderModel = new OrderModel();
-
-            orderModel.CartModel = new CartModel()
+            model.CartModel = new CartModel()
             {
                 CartId = cart.Id,
                 CartItems = cart.CartItems.Select(i => new CartItemModel()
@@ -74,208 +123,190 @@ namespace ShopApp.WebUI.Controllers
                     CartItemId = i.Id,
                     ProductId = i.Product.Id,
                     Name = i.Product.Name,
-                    Price = (decimal)i.Product.Price,
+                    Price = i.Product.Price ?? 0,
                     ImageUrl = i.Product.ImageUrl,
                     Quantity = i.Quantity
                 }).ToList()
             };
 
-            return View(orderModel);
-        }
+            var payment = PaymentProcess(model);
 
-        [HttpPost]
-        public IActionResult Checkout(OrderModel model)
-        {
-            if (ModelState.IsValid)                                                // ***************************
+            if (payment.Status == "success")
             {
-                var userId = _userManager.GetUserId(User);
-                var cart = _cartService.GetCartByUserId(userId);
-
-                model.CartModel = new CartModel()
-                {
-                    CartId = cart.Id,
-                    CartItems = cart.CartItems.Select(i => new CartItemModel()
-                    {
-                        CartItemId = i.Id,
-                        ProductId = i.Product.Id,
-                        Name = i.Product.Name,
-                        Price = (decimal)i.Product.Price,
-                        ImageUrl = i.Product.ImageUrl,
-                        Quantity = i.Quantity
-                    }).ToList()
-                };
-
-                var payment = PaymentProcess(model);                                  //*****************
-
-                if (payment.Status == "success")
-                {
-                    SaveOrder(model, payment, userId); //11111111111
-                    ClearCart(cart.Id.ToString());
-                    return View("Success");
-                }
+                SaveOrder(model, payment, userId);
+                ClearCart(cart.Id.ToString());
+                return View("Success");
             }
-
-            return View(model);
         }
 
-        private void SaveOrder(OrderModel model, Payment payment, string userId)
+        return View(model);
+    }
+
+    private void SaveOrder(OrderModel model, Payment payment, string userId)
+    {
+        var order = new Order
         {
-            var order = new Order();
+            OrderNumber = new Random().Next(111111, 999999).ToString(),
+            OrderState = EnumOrderState.Completed,
+            PaymentTypes = EnumPaymentTypes.CreditCart,
+            PaymentId = payment.PaymentId,
+            ConversationId = payment.ConversationId,
+            OrderDate = DateTime.Now,
+            FirstName = model.FirstName ?? string.Empty,
+            LastName = model.LastName ?? string.Empty,
+            Email = model.Email,
+            Phone = model.Phone,
+            Address = model.Address,
+            UserId = userId
+        };
 
-            order.OrderNumber = new Random().Next(111111, 999999).ToString();
-            order.OrderState = EnumOrderState.Completed;
-            order.PaymentTypes = EnumPaymentTypes.CreditCart;
-            order.PaymentId = payment.PaymentId;
-            order.ConversationId = payment.ConversationId;
-            order.OrderDate = new DateTime();
-            order.FirstName = model.FirstName;
-            order.LastName = model.LastName;
-            order.Email = model.Email;
-            order.Phone = model.Phone;
-            order.Address = model.Address;
-            order.UserId = userId;
-
+        if (model.CartModel?.CartItems != null)
+        {
             foreach (var item in model.CartModel.CartItems)
             {
-                var orderitem = new OrderItem()
+                var orderitem = new ShopApp.Entities.OrderItem()
                 {
                     Price = item.Price,
                     Quantity = item.Quantity,
                     ProductId = item.ProductId
                 };
-                order.OrderItems.Add(orderitem);     //2222222222222222222222222
+                order.OrderItems.Add(orderitem);
             }
-            _orderService.Create(order);
         }
+        _orderService.Create(order);
+    }
 
-        private void ClearCart(string cartId)
+    private void ClearCart(string cartId)
+    {
+        _cartService.ClearCart(cartId);
+    }
+
+    private Payment PaymentProcess(OrderModel model)
+    {
+        var options = new Options
         {
-            _cartService.ClearCart(cartId);
-        }
+            ApiKey = _configuration["Iyzico:ApiKey"] ?? "sandbox-ECA09KNTaes3IetV1JpkKpO1nL3pKrEo",
+            SecretKey = _configuration["Iyzico:SecretKey"] ?? "sandbox-wDxAvfX3zjyNK0ZJvU7MFGl6zCcWSoB3",
+            BaseUrl = "https://sandbox-api.iyzipay.com"
+        };
 
-        private Payment PaymentProcess(OrderModel model)
+        var request = new CreatePaymentRequest
         {
-            Options options = new Options();
-            options.ApiKey = "sandbox-ECA09KNTaes3IetV1JpkKpO1nL3pKrEo";                      //iyzico api bölgesi***********
-            options.SecretKey = "sandbox-wDxAvfX3zjyNK0ZJvU7MFGl6zCcWSoB3";
-            options.BaseUrl = "https://sandbox-api.iyzipay.com";
+            Locale = Locale.TR.ToString(),
+            ConversationId = Guid.NewGuid().ToString(),
+            Price = model.CartModel?.TotalPrice().ToString("F0") ?? "0",
+            PaidPrice = model.CartModel?.TotalPrice().ToString("F0") ?? "0",
+            Currency = Currency.TRY.ToString(),
+            Installment = 1,
+            BasketId = model.CartModel?.CartId.ToString() ?? "0",
+            PaymentChannel = PaymentChannel.WEB.ToString(),
+            PaymentGroup = PaymentGroup.PRODUCT.ToString()
+        };
 
-            CreatePaymentRequest request = new CreatePaymentRequest();
-            request.Locale = Locale.TR.ToString();
-            request.ConversationId = Guid.NewGuid().ToString();
-            request.Price = model.CartModel.TotalPrice().ToString().Split(",")[0]; ;
-            request.PaidPrice = model.CartModel.TotalPrice().ToString().Split(",")[0]; ;
-            request.Currency = Currency.TRY.ToString();
-            request.Installment = 1;
-            request.BasketId = model.CartModel.CartId.ToString();
-            request.PaymentChannel = PaymentChannel.WEB.ToString();
-            request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+        var paymentCard = new PaymentCard
+        {
+            CardHolderName = model.CardName,
+            CardNumber = model.CardNumber,
+            ExpireMonth = model.ExpirationMonth,
+            ExpireYear = model.ExpirationYear,
+            Cvc = model.Cvv,
+            RegisterCard = 0
+        };
+        request.PaymentCard = paymentCard;
 
-            PaymentCard paymentCard = new PaymentCard();
-            paymentCard.CardHolderName = model.CardName;
-            paymentCard.CardNumber = model.CardNumber;
-            paymentCard.ExpireMonth = model.ExpirationMonth;
-            paymentCard.ExpireYear = model.ExpirationYear;
-            paymentCard.Cvc = model.Cvv;
-            paymentCard.RegisterCard = 0;
-            request.PaymentCard = paymentCard;
+        var buyer = new Buyer
+        {
+            Id = "BY789",
+            Name = "John",
+            Surname = "Doe",
+            GsmNumber = "+905350000000",
+            Email = "email@email.com",
+            IdentityNumber = "74300864791",
+            LastLoginDate = "2015-10-05 12:43:35",
+            RegistrationDate = "2013-04-21 15:12:09",
+            RegistrationAddress = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1",
+            Ip = "85.34.78.112",
+            City = "Istanbul",
+            Country = "Turkey",
+            ZipCode = "34732"
+        };
+        request.Buyer = buyer;
 
-            //paymentCard.CardHolderName = "John Doe";
-            //paymentCard.CardNumber = "5528790000000008";
-            //paymentCard.ExpireMonth = "12";
-            //paymentCard.ExpireYear = "2030";
-            //paymentCard.Cvc = "123";
+        var shippingAddress = new Address
+        {
+            ContactName = "Jane Doe",
+            City = "Istanbul",
+            Country = "Turkey",
+            Description = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1",
+            ZipCode = "34742"
+        };
+        request.ShippingAddress = shippingAddress;
 
-            Buyer buyer = new Buyer();
-            buyer.Id = "BY789";
-            buyer.Name = "John";
-            buyer.Surname = "Doe";
-            buyer.GsmNumber = "+905350000000";
-            buyer.Email = "email@email.com";
-            buyer.IdentityNumber = "74300864791";
-            buyer.LastLoginDate = "2015-10-05 12:43:35";
-            buyer.RegistrationDate = "2013-04-21 15:12:09";
-            buyer.RegistrationAddress = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
-            buyer.Ip = "85.34.78.112";
-            buyer.City = "Istanbul";
-            buyer.Country = "Turkey";
-            buyer.ZipCode = "34732";
-            request.Buyer = buyer;
+        var billingAddress = new Address
+        {
+            ContactName = "Jane Doe",
+            City = "Istanbul",
+            Country = "Turkey",
+            Description = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1",
+            ZipCode = "34742"
+        };
+        request.BillingAddress = billingAddress;
 
-            Address shippingAddress = new Address();
-            shippingAddress.ContactName = "Jane Doe";
-            shippingAddress.City = "Istanbul";
-            shippingAddress.Country = "Turkey";
-            shippingAddress.Description = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
-            shippingAddress.ZipCode = "34742";
-            request.ShippingAddress = shippingAddress;
+        var basketItems = new List<BasketItem>();
 
-            Address billingAddress = new Address();
-            billingAddress.ContactName = "Jane Doe";
-            billingAddress.City = "Istanbul";
-            billingAddress.Country = "Turkey";
-            billingAddress.Description = "Nidakule Göztepe, Merdivenköy Mah. Bora Sok. No:1";
-            billingAddress.ZipCode = "34742";
-            request.BillingAddress = billingAddress;
-
-            List<BasketItem> basketItems = new List<BasketItem>();
-            BasketItem basketItem;
-
+        if (model.CartModel?.CartItems != null)
+        {
             foreach (var item in model.CartModel.CartItems)
             {
-                basketItem = new BasketItem();
-                basketItem.Id = item.ProductId.ToString();
-                basketItem.Name = item.Name;
-                basketItem.Category1 = "Sebze";
-                basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
-                // basketItem.Price = item.Price.ToString().Split(",")[0];  // hata düzelt***************************************************
-                // basketItem.Price = (item.Quantity*item.Price).ToString().Split(",")[0];
-                basketItem.Price = (item.Quantity * item.Price).ToString().Split(",")[0];
-
+                var basketItem = new BasketItem
+                {
+                    Id = item.ProductId.ToString(),
+                    Name = item.Name,
+                    Category1 = "Sebze",
+                    ItemType = BasketItemType.PHYSICAL.ToString(),
+                    Price = (item.Quantity * item.Price).ToString("F0")
+                };
                 basketItems.Add(basketItem);
             }
-
-            request.BasketItems = basketItems;
-
-            return Payment.Create(request, options);
-
         }
-        public IActionResult GetOrders()
+
+        request.BasketItems = basketItems;
+
+        return Payment.Create(request, options);
+    }
+
+    public IActionResult GetOrders()
+    {
+        var orders = _orderService.GetOrders(null);
+        var orderListModel = new List<OrderListModel>();
+
+        foreach (var order in orders)
         {
-            var orders = _orderService.GetOrders(null);// var orders = _orderService.GetOrders(_userManager.GetUserId(User));
-            var orderListModel = new List<OrderListModel>();
-            OrderListModel orderModel;
-
-            foreach (var order in orders)
+            var orderModel = new OrderListModel
             {
-                orderModel = new OrderListModel();
-                orderModel.OrderId = order.Id;
-                orderModel.OrderNumber = order.OrderNumber;
-                orderModel.OrderDate = order.OrderDate;
-                orderModel.OrderNote = order.OrderNote;
-                orderModel.Phone = order.Phone;
-                orderModel.FirstName = order.FirstName;
-                orderModel.LastName = order.LastName;
-                orderModel.Email = order.Email;
-                orderModel.Address = order.Address;
-                orderModel.City = order.City;
-
-                orderModel.OrderItems = order.OrderItems.Select(i => new OrderItemModel()
+                OrderId = order.Id,
+                OrderNumber = order.OrderNumber,
+                OrderDate = order.OrderDate,
+                OrderNote = order.OrderNote,
+                Phone = order.Phone,
+                FirstName = order.FirstName,
+                LastName = order.LastName,
+                Email = order.Email,
+                Address = order.Address,
+                City = order.City,
+                OrderItems = order.OrderItems.Select(i => new OrderItemModel()
                 {
                     OrderItemId = i.Id,
                     Name = i.Product.Name,
                     Price = i.Price,
                     Quantity = i.Quantity,
                     ImageUrl = i.Product.ImageUrl
-                }).ToList();
+                }).ToList()
+            };
 
-                orderListModel.Add(orderModel);
-
-            }
-
-            return View(orderListModel);
+            orderListModel.Add(orderModel);
         }
 
-
+        return View(orderListModel);
     }
 }
