@@ -1,26 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using ShopApp.Business.Abstract;
 using ShopApp.WebUI.Models;
 
-namespace ShopApp.WebUI.Controllers
+namespace ShopApp.WebUI.Controllers;
+
+public class HomeController : Controller
 {
-    public class HomeController : Controller
+    private readonly IProductService _productService;
+
+    public HomeController(IProductService productService)
     {
-        private IProductService _productService;
-        public HomeController(IProductService productService)
+        _productService = productService;
+    }
+
+    public IActionResult Index()
+    {
+        return View(new ProductListModel()
         {
-            _productService = productService;
-        }
-        public IActionResult Index()
-        {
-            return View(new ProductListModel()
-            {
-                Products = _productService.GetAll()
-            });
-        }
+            Products = _productService.GetAll()
+        });
     }
 }
