@@ -1,24 +1,18 @@
 ﻿using Microsoft.AspNetCore.Mvc.ViewFeatures;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using System.Text.Json;
 
-namespace ShopApp.WebUI.Extensions
+namespace ShopApp.WebUI.Extensions;
+
+public static class TempDataExtensions
 {
-    public static class TempDataExtensions
+    public static void Put<T>(this ITempDataDictionary tempData, string key, T value) where T : class
     {
-        public static void Put<T>(this ITempDataDictionary tempData, string key, T value) where T : class
-        {
-            tempData[key] = JsonConvert.SerializeObject(value);
-        }
+        tempData[key] = JsonSerializer.Serialize(value);
+    }
 
-        public static T Get<T>(this ITempDataDictionary tempData, string key) where T : class
-        {
-            object o;
-            tempData.TryGetValue(key, out o);
-            return o == null ? null : JsonConvert.DeserializeObject<T>((string)o);
-        }
+    public static T? Get<T>(this ITempDataDictionary tempData, string key) where T : class
+    {
+        tempData.TryGetValue(key, out object? o);
+        return o == null ? null : JsonSerializer.Deserialize<T>((string)o);
     }
 }
