@@ -1,17 +1,11 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.Identity
+namespace ShopApp.WebUI.Identity;
+
+public class ApplicationIdentityDbContext : IdentityDbContext<ApplicationUser>
 {
-    public class ApplicationIdentityDbContext:IdentityDbContext<ApplicationUser>
+    public ApplicationIdentityDbContext(DbContextOptions<ApplicationIdentityDbContext> options) : base(options)
     {
-        public ApplicationIdentityDbContext(DbContextOptions<ApplicationIdentityDbContext> options) : base(options)
-        {
-
-        }
     }
 }
