@@ -1,13 +1,8 @@
 ﻿using Microsoft.AspNetCore.Identity;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.Identity
+namespace ShopApp.WebUI.Identity;
+
+public class ApplicationUser : IdentityUser
 {
-    public class ApplicationUser:IdentityUser
-    {
-        public string FullName { get; set; }
-    }
+    public string FullName { get; set; } = string.Empty;
 }
