@@ -1,25 +1,23 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.FileProviders;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Microsoft.Extensions.FileProviders;
 
-namespace ShopApp.WebUI.Middlewares
+namespace ShopApp.WebUI.Middlewares;
+
+public static class ApplicationBuilderExtensions
 {
-    public static class ApplicationBuilderExtensions
+    public static IApplicationBuilder CustomStaticfiles(this IApplicationBuilder app)
     {
-        public static IApplicationBuilder CustomStaticfiles(this IApplicationBuilder app)
+        var path = Path.Combine(Directory.GetCurrentDirectory(), "node_modules");
+        
+        if (Directory.Exists(path))
         {
-            var path = Path.Combine(Directory.GetCurrentDirectory(), "node_modules");
             var options = new StaticFileOptions
             {
                 FileProvider = new PhysicalFileProvider(path),
                 RequestPath = "/modules"
             };
             app.UseStaticFiles(options);
-            return app;
         }
+        
+        return app;
     }
 }
