@@ -1,13 +1,8 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class CategoryListModel
 {
-    public class CategoryListModel
-    {
-        public List<Category> Categories { get; set; }
-    }
+    public List<Category> Categories { get; set; } = new();
 }
