@@ -1,14 +1,9 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class CategoryListViewModel
 {
-    public class CategoryListViewModel
-    {
-        public string SelectedCategory { get; set; }
-        public List<Category> Categories { get; set; }
-    }
+    public string? SelectedCategory { get; set; }
+    public List<Category> Categories { get; set; } = new();
 }
