@@ -1,20 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class LoginModel
 {
-    public class LoginModel
-    {
-        [Required(ErrorMessage = "kullanıcı adı belirtiniz")]
-        [DataType(DataType.EmailAddress)]
-        public string Email { get; set; }
+    [Required(ErrorMessage = "kullanıcı adı belirtiniz")]
+    [DataType(DataType.EmailAddress)]
+    public string? Email { get; set; }
 
-        [Required(ErrorMessage = "parola belirtiniz")]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-        public string ReturnUrl { get; set; }
-    }
+    [Required(ErrorMessage = "parola belirtiniz")]
+    [DataType(DataType.Password)]
+    public string? Password { get; set; }
+
+    public string? ReturnUrl { get; set; }
 }
