@@ -1,41 +1,36 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class OrderListModel
 {
-    public class OrderListModel
+    public int OrderId { get; set; }
+    public string OrderNumber { get; set; } = string.Empty;
+    public DateTime OrderDate { get; set; }
+    public EnumOrderState OrderState { get; set; }
+    public EnumPaymentTypes PaymentTypes { get; set; }
+
+    public string FirstName { get; set; } = string.Empty;
+    public string LastName { get; set; } = string.Empty;
+    public string? Address { get; set; }
+    public string? City { get; set; }
+    public string? Phone { get; set; }
+    public string? Email { get; set; }
+    public string? OrderNote { get; set; }
+
+    public List<OrderItemModel> OrderItems { get; set; } = new();
+
+    public decimal TotalPrice()
     {
-        public int OrderId { get; set; }
-        public string OrderNumber { get; set; }
-        public DateTime OrderDate { get; set; }
-        public EnumOrderState OrderState { get; set; }
-        public EnumPaymentTypes PaymentTypes { get; set; }
-
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Address { get; set; }
-        public string City { get; set; }
-        public string Phone { get; set; }
-        public string Email { get; set; }
-        public string OrderNote { get; set; }
-
-        public List<OrderItemModel> OrderItems { get; set; }
-
-        public decimal TotalPrice()
-        {
-            return OrderItems.Sum(i => i.Price * i.Quantity);
-        }
+        return OrderItems.Sum(i => i.Price * i.Quantity);
     }
+}
 
-    public class OrderItemModel
-    {
-        public int OrderItemId { get; set; }
-        public decimal Price { get; set; }
-        public int Quantity { get; set; }
-        public string Name { get; set; }
-        public string ImageUrl { get; set; }
-    }
+public class OrderItemModel
+{
+    public int OrderItemId { get; set; }
+    public decimal Price { get; set; }
+    public int Quantity { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string? ImageUrl { get; set; }
 }
