@@ -1,25 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class OrderModel
 {
-    public class OrderModel
-    {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Address { get; set; }
-        [Display(Name = "Şehir")]
-        public string City { get; set; }
-        public string Phone { get; set; }
-        public string Email { get; set; }
-        public string CardName { get; set; }
-        public string CardNumber { get; set; }
-        public string ExpirationMonth { get; set; }
-        public string ExpirationYear { get; set; }
-        public string Cvv { get; set; }
-        public CartModel CartModel { get; set; }
-    }
+    public string? FirstName { get; set; }
+    public string? LastName { get; set; }
+    public string? Address { get; set; }
+
+    [Display(Name = "Şehir")]
+    public string? City { get; set; }
+    public string? Phone { get; set; }
+    public string? Email { get; set; }
+    public string? CardName { get; set; }
+    public string? CardNumber { get; set; }
+    public string? ExpirationMonth { get; set; }
+    public string? ExpirationYear { get; set; }
+    public string? Cvv { get; set; }
+    public CartModel? CartModel { get; set; }
 }
