@@ -1,27 +1,22 @@
 ﻿using ShopApp.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class PageInfo
 {
-    public class PageInfo
-    {
-        public int TotalItems { get; set; }
-        public int ItemsPerPage { get; set; }
-        public int CurrentPage { get; set; }
-        public string CurrentCategory { get; set; }
-        public int TotalPages()
-        {
-            return (int)Math.Ceiling((decimal)TotalItems / ItemsPerPage);
-        }
+    public int TotalItems { get; set; }
+    public int ItemsPerPage { get; set; }
+    public int CurrentPage { get; set; }
+    public string? CurrentCategory { get; set; }
 
-    }
-    public class ProductListModel
+    public int TotalPages()
     {
-        public PageInfo PageInfo { get; set; }
-        public List<Product> Products { get; set; }
-        
+        return (int)Math.Ceiling((decimal)TotalItems / ItemsPerPage);
     }
+}
+
+public class ProductListModel
+{
+    public PageInfo? PageInfo { get; set; }
+    public List<Product> Products { get; set; } = new();
 }
