@@ -1,20 +1,17 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace ShopApp.WebUI.Models
+namespace ShopApp.WebUI.Models;
+
+public class ResetPasswordModel
 {
-    public class ResetPasswordModel
-    {
-        [Required]
-        public string Token { get; set; }
-        [Required]
-        [DataType(DataType.EmailAddress)]
-        public string Email { get; set; }
-        [Required]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-    }
+    [Required]
+    public string? Token { get; set; }
+
+    [Required]
+    [DataType(DataType.EmailAddress)]
+    public string? Email { get; set; }
+
+    [Required]
+    [DataType(DataType.Password)]
+    public string? Password { get; set; }
 }
