@@ -1,27 +1,24 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ShopApp.Business.Abstract;
 using ShopApp.WebUI.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace ShopApp.WebUI.ViewComponents
+namespace ShopApp.WebUI.ViewComponents;
+
+public class CategoryListViewComponent : ViewComponent
 {
-    public class CategoryListViewComponent:ViewComponent
+    private readonly ICategoryService _categoryService;
+
+    public CategoryListViewComponent(ICategoryService categoryService)
     {
-        private ICategoryService _categoryService;
-        public CategoryListViewComponent(ICategoryService categoryService)
+        _categoryService = categoryService;
+    }
+
+    public IViewComponentResult Invoke()
+    {
+        return View(new CategoryListViewModel()
         {
-            _categoryService = categoryService;
-        }
-        public IViewComponentResult Invoke()
-        {
-            return View(new CategoryListViewModel()
-            {
-                SelectedCategory = RouteData.Values["category"]?.ToString(),
-                Categories = _categoryService.GetAll()
-            }) ;
-        }
+            SelectedCategory = RouteData.Values["category"]?.ToString(),
+            Categories = _categoryService.GetAll()
+        });
     }
 }
