@@ -1,12 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace ShopApp.Business.Abstract;
 
-namespace ShopApp.Business.Abstract
+public interface IValidator<T>
 {
-   public interface IValidator<T>
-    {
-        string ErrorMessage { get; set; }
-        bool Validate(T entity);
-    }
+    string ErrorMessage { get; set; }
+    bool Validate(T entity);
 }
