@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace ShopApp.Entities;
 
-namespace ShopApp.Entities
+public class Cart
 {
-   public class Cart
-    {
-        public int Id { get; set; }
-        public string UserId { get; set; }
-
-        public List<CartItem> CartItems { get; set; }
-    }
+    public int Id { get; set; }
+    public string UserId { get; set; } = string.Empty;
+    public List<CartItem> CartItems { get; set; } = new();
 }
