@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace ShopApp.Entities;
 
-namespace ShopApp.Entities
+public class Category
 {
-    public class Category
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public List<ProductCategory> ProductCategories { get; set; }
-
-    }
+    public int Id { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public List<ProductCategory> ProductCategories { get; set; } = new();
 }
