@@ -1,20 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace ShopApp.Entities;
 
-namespace ShopApp.Entities
+public class OrderItem
 {
-    public class OrderItem
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public int OrderId { get; set; }
-        public Order Order { get; set; }
+    public int OrderId { get; set; }
+    public Order Order { get; set; } = null!;
 
-        public int ProductId { get; set; }
-        public Product Product { get; set; }
+    public int ProductId { get; set; }
+    public Product Product { get; set; } = null!;
 
-        public decimal Price { get; set; }
-        public int Quantity { get; set; }
-    }
+    public decimal Price { get; set; }
+    public int Quantity { get; set; }
 }
