@@ -1,29 +1,22 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace ShopApp.WebUI.Models;
 
-namespace ShopApp.WebUI.Models
+public class CartModel
 {
-    public class CartModel
+    public int CartId { get; set; }
+    public List<CartItemModel> CartItems { get; set; } = new();
+
+    public decimal TotalPrice()
     {
-        public int CartId { get; set; }
-        public List<CartItemModel> CartItems { get; set; }
-
-        public decimal TotalPrice()
-        {
-            return CartItems.Sum(i => i.Price * i.Quantity);
-        }
+        return CartItems.Sum(i => i.Price * i.Quantity);
     }
+}
 
-    public class CartItemModel
-    {
-        public int CartItemId { get; set; }
-
-        public int ProductId { get; set; }
-        public string Name { get; set; }
-        public decimal Price { get; set; }
-        public string ImageUrl { get; set; }
-        public int Quantity { get; set; }
-    }
+public class CartItemModel
+{
+    public int CartItemId { get; set; }
+    public int ProductId { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public decimal Price { get; set; }
+    public string? ImageUrl { get; set; }
+    public int Quantity { get; set; }
 }
