@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿namespace ShopApp.WebUI.Models;
 
-namespace ShopApp.WebUI.Models
+public class ResultMessage
 {
-    public class ResultMessage
-    {
-        public string Title { get; set; }
-        public string Message { get; set; }
-        public string Css { get; set; }
-    }
+    public string Title { get; set; } = string.Empty;
+    public string Message { get; set; } = string.Empty;
+    public string Css { get; set; } = string.Empty;
 }
