@@ -1,52 +1,60 @@
 ﻿@model LoginModel
 @{
-    ViewData["Title"] = "Login";
+    ViewData["Title"] = "Giriş Yap";
 }
 
-@section Scripts {
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation\dist\jquery.validate.min.js"></script>
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation-unobtrusive\dist\jquery.validate.unobtrusive.min.js"></script>
-}
+<div class="row justify-content-center">
+    <div class="col-md-6 col-lg-4">
+        <div class="card shadow">
+            <div class="card-header bg-primary text-white text-center">
+                <h4 class="mb-0"><i class="bi bi-box-arrow-in-right me-2"></i>Giriş Yap</h4>
+            </div>
+            <div class="card-body">
+                @if (ViewBag.ErrorMessage != null)
+                {
+                    <div class="alert alert-danger">@ViewBag.ErrorMessage</div>
+                }
 
-<h1>Login</h1>
-<hr />
+                <form asp-controller="Account" asp-action="Login" method="post">
+                    <div asp-validation-summary="All" class="text-danger mb-3"></div>
+                    <input type="hidden" name="ReturnUrl" value="@Model.ReturnUrl" />
+                    
+                    <div class="mb-3">
+                        <label asp-for="Email" class="form-label">E-posta</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-envelope"></i></span>
+                            <input asp-for="Email" class="form-control" placeholder="ornek@email.com" />
+                        </div>
+                        <span asp-validation-for="Email" class="text-danger small"></span>
+                    </div>
 
-<div class="row">
-    <div class="col-md-8">
+                    <div class="mb-3">
+                        <label asp-for="Password" class="form-label">Şifre</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-lock"></i></span>
+                            <input asp-for="Password" class="form-control" placeholder="••••••••" />
+                        </div>
+                        <span asp-validation-for="Password" class="text-danger small"></span>
+                    </div>
 
-        @if (ViewBag.ErrorMessage != null)
-        {
-            <div class="text-danger">@ViewBag.ErrorMessage</div>
-        }
+                    <div class="d-grid gap-2">
+                        <button type="submit" class="btn btn-primary">
+                            <i class="bi bi-box-arrow-in-right me-2"></i>Giriş Yap
+                        </button>
+                    </div>
+                </form>
 
-        <form asp-controller="Account" asp-action="Login" method="post">
-            <div asp-validation-summary="All" class="text-danger"></div>
-            <input type="hidden" name="ReturnUrl" value="@Model.ReturnUrl" />
-            <div class="form-group row">
-                <label asp-for="Email" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Email" value="@Model.Email" class="form-control" />
-                    <span asp-validation-for="Email" class="text-danger"></span>
+                <hr />
+                <div class="text-center">
+                    <a asp-controller="Account" asp-action="ForgotPassword" class="text-decoration-none">
+                        Şifremi Unuttum
+                    </a>
+                    <span class="mx-2">|</span>
+                    <a asp-controller="Account" asp-action="Register" class="text-decoration-none">
+                        Hesap Oluştur
+                    </a>
                 </div>
             </div>
-
-            <div class="form-group row">
-                <label asp-for="Password" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Password" value="@Model.Password" class="form-control" />
-                    <span asp-validation-for="Password" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <div class="col-md-10 offset-md-2">
-                    <button type="submit" class="btn btn-primary">Login</button>
-                </div>
-            </div>
-
-        </form>
-
+        </div>
     </div>
 </div>
-
-
