@@ -1,72 +1,83 @@
 ﻿@model RegisterModel
 @{
-    ViewData["Title"] = "Register";
+    ViewData["Title"] = "Kayıt Ol";
 }
 
-@section Scripts {
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation\dist\jquery.validate.min.js"></script>
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation-unobtrusive\dist\jquery.validate.unobtrusive.min.js"></script>
-}
+<div class="row justify-content-center">
+    <div class="col-md-6 col-lg-5">
+        <div class="card shadow">
+            <div class="card-header bg-primary text-white text-center">
+                <h4 class="mb-0"><i class="bi bi-person-plus me-2"></i>Hesap Oluştur</h4>
+            </div>
+            <div class="card-body">
+                @if (ViewBag.ErrorMessage != null)
+                {
+                    <div class="alert alert-danger">@ViewBag.ErrorMessage</div>
+                }
 
-<h1>Kullanıcı Üyeliği</h1>
-<hr />
+                <form asp-controller="Account" asp-action="Register" method="post">
+                    <div asp-validation-summary="All" class="text-danger mb-3"></div>
+                    
+                    <div class="mb-3">
+                        <label asp-for="FullName" class="form-label">Ad Soyad</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-person"></i></span>
+                            <input asp-for="FullName" class="form-control" placeholder="Adınız Soyadınız" />
+                        </div>
+                        <span asp-validation-for="FullName" class="text-danger small"></span>
+                    </div>
 
-<div class="row">
-    <div class="col-md-8">
+                    <div class="mb-3">
+                        <label asp-for="Username" class="form-label">Kullanıcı Adı</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-at"></i></span>
+                            <input asp-for="Username" class="form-control" placeholder="kullanıcıadı" />
+                        </div>
+                        <span asp-validation-for="Username" class="text-danger small"></span>
+                    </div>
 
-        @if (ViewBag.ErrorMessage != null)
-        {
-            <div class="text-danger">@ViewBag.ErrorMessage</div>
-        }
+                    <div class="mb-3">
+                        <label asp-for="Email" class="form-label">E-posta</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-envelope"></i></span>
+                            <input asp-for="Email" class="form-control" placeholder="ornek@email.com" />
+                        </div>
+                        <span asp-validation-for="Email" class="text-danger small"></span>
+                    </div>
 
-        <form asp-controller="Account" asp-action="Register" method="post">
-            <div asp-validation-summary="All" class="text-danger"></div>
-            <div class="form-group row">
-                <label asp-for="FullName" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="FullName" value="@Model.FullName" class="form-control" />
-                    <span asp-validation-for="FullName" class="text-danger"></span>
+                    <div class="mb-3">
+                        <label asp-for="Password" class="form-label">Şifre</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-lock"></i></span>
+                            <input asp-for="Password" class="form-control" placeholder="••••••••" />
+                        </div>
+                        <span asp-validation-for="Password" class="text-danger small"></span>
+                    </div>
+
+                    <div class="mb-3">
+                        <label asp-for="RePassword" class="form-label">Şifre Tekrar</label>
+                        <div class="input-group">
+                            <span class="input-group-text"><i class="bi bi-lock-fill"></i></span>
+                            <input asp-for="RePassword" class="form-control" placeholder="••••••••" />
+                        </div>
+                        <span asp-validation-for="RePassword" class="text-danger small"></span>
+                    </div>
+
+                    <div class="d-grid gap-2">
+                        <button type="submit" class="btn btn-primary">
+                            <i class="bi bi-person-plus me-2"></i>Hesap Oluştur
+                        </button>
+                    </div>
+                </form>
+
+                <hr />
+                <div class="text-center">
+                    <span>Zaten hesabınız var mı?</span>
+                    <a asp-controller="Account" asp-action="Login" class="text-decoration-none ms-2">
+                        Giriş Yap
+                    </a>
                 </div>
             </div>
-            <div class="form-group row">
-                <label asp-for="Username" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Username" value="@Model.Username" class="form-control" />
-                    <span asp-validation-for="Username" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label asp-for="Email" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Email" value="@Model.Email" class="form-control" />
-                    <span asp-validation-for="Email" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label asp-for="Password" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Password" value="@Model.Password" class="form-control" />
-                    <span asp-validation-for="Password" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <label asp-for="RePassword" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="RePassword" value="@Model.RePassword" class="form-control" />
-                    <span asp-validation-for="RePassword" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <div class="col-md-10 offset-md-2">
-                    <button type="submit" class="btn btn-primary">Create User</button>
-                </div>
-            </div>
-
-        </form>
-
+        </div>
     </div>
 </div>
