@@ -1,36 +1,43 @@
 ﻿@model CategoryModel
 @{
-    ViewData["Title"] = "EditCategory";
+    ViewData["Title"] = "Yeni Kategori Ekle";
 }
-@section Scripts {
 
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation\dist\jquery.validate.min.js"></script>
-
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation-unobtrusive\dist\jquery.validate.unobtrusive.min.js"></script>
-}
-<h1>Create Category</h1>
-<hr />
+<div class="d-flex justify-content-between align-items-center mb-4">
+    <h1 class="h3 mb-0 text-gray-800">Yeni Kategori Ekle</h1>
+    <a href="/admin/categories" class="btn btn-outline-secondary btn-sm">
+        <i class="bi bi-arrow-left"></i> Geri Dön
+    </a>
+</div>
 
 <div class="row">
     <div class="col-md-8">
-        
+        <div class="card shadow mb-4">
+            <div class="card-header py-3">
+                <h6 class="m-0 font-weight-bold text-primary">Kategori Bilgileri</h6>
+            </div>
+            <div class="card-body">
                 <form asp-controller="Admin" asp-action="CreateCategory" method="post">
-                    <div asp-validation-summary="All" class="text-danger"></div>
-                    <div class="form-group row">
-                        <label asp-for="Name" class="col-md-2 col-form-label">isim</label>
-                        <div class="col-md-10">
-                            <input asp-for="Name" value="" class="form-control" />
-                            <span asp-validation-for="Name" class="text-danger"></span>
+                    <div asp-validation-summary="All" class="text-danger mb-3"></div>
+                    
+                    <div class="mb-3 row">
+                        <label asp-for="Name" class="col-sm-3 col-form-label text-end fw-bold">Kategori Adı</label>
+                        <div class="col-sm-9">
+                            <input asp-for="Name" class="form-control" placeholder="Kategori adı giriniz" />
+                            <span asp-validation-for="Name" class="text-danger small"></span>
                         </div>
                     </div>
 
-                    <div class="form-group row">
-                        <div class="col-md-10 offset-md-2">
-                            <button type="submit" class="btn btn-primary">Save Category</button>
+                    <div class="row">
+                        <div class="col-sm-9 offset-sm-3">
+                            <button type="submit" class="btn btn-primary">
+                                <i class="bi bi-save me-2"></i> Kaydet
+                            </button>
                         </div>
                     </div>
 
                 </form>
-
             </div>
+        </div>
     </div>
+</div>
