@@ -1,111 +1,98 @@
 ﻿@model ProductModel
 @{
-    ViewData["Title"] = "CreateProduct";
+    ViewData["Title"] = "Yeni Ürün Ekle";
 }
+
+<div class="d-flex justify-content-between align-items-center mb-4">
+    <h1 class="h3 mb-0 text-gray-800">Yeni Ürün Ekle</h1>
+    <a href="/admin/products" class="btn btn-outline-secondary btn-sm">
+        <i class="bi bi-arrow-left"></i> Geri Dön
+    </a>
+</div>
+
+<div class="row">
+    <div class="col-lg-8">
+        <div class="card shadow mb-4">
+            <div class="card-header py-3">
+                <h6 class="m-0 font-weight-bold text-primary">Ürün Bilgileri</h6>
+            </div>
+            <div class="card-body">
+                @if (ViewBag.ErrorMessage != null)
+                {
+                    <div class="alert alert-danger">@ViewBag.ErrorMessage</div>
+                }
+
+                <form asp-controller="Admin" asp-action="CreateProduct" method="post">
+                    <div asp-validation-summary="All" class="text-danger mb-3"></div>
+                    
+                    <div class="mb-3 row">
+                        <label asp-for="Name" class="col-sm-3 col-form-label text-end fw-bold">Ürün Adı</label>
+                        <div class="col-sm-9">
+                            <input asp-for="Name" class="form-control" placeholder="Ürün adını giriniz" />
+                            <span asp-validation-for="Name" class="text-danger small"></span>
+                        </div>
+                    </div>
+
+                    <div class="mb-3 row">
+                        <label asp-for="ImageUrl" class="col-sm-3 col-form-label text-end fw-bold">Görsel URL</label>
+                        <div class="col-sm-9">
+                            <input asp-for="ImageUrl" class="form-control" placeholder="ornek.jpg" />
+                            <span asp-validation-for="ImageUrl" class="text-danger small"></span>
+                        </div>
+                    </div>
+
+                    <div class="mb-3 row">
+                        <label asp-for="Price" class="col-sm-3 col-form-label text-end fw-bold">Fiyat</label>
+                        <div class="col-sm-9">
+                            <div class="input-group">
+                                <input asp-for="Price" class="form-control" placeholder="0.00" type="number" step="0.01" />
+                                <span class="input-group-text">₺</span>
+                            </div>
+                            <span asp-validation-for="Price" class="text-danger small"></span>
+                        </div>
+                    </div>
+
+                    <div class="mb-3 row">
+                        <label asp-for="Description" class="col-sm-3 col-form-label text-end fw-bold">Açıklama</label>
+                        <div class="col-sm-9">
+                            <textarea id="editor" asp-for="Description" class="form-control"></textarea>
+                            <span asp-validation-for="Description" class="text-danger small"></span>
+                        </div>
+                    </div>
+
+                    <div class="row">
+                        <div class="col-sm-9 offset-sm-3">
+                            <button type="submit" class="btn btn-primary">
+                                <i class="bi bi-save me-2"></i> Ürünü Kaydet
+                            </button>
+                        </div>
+                    </div>
+
+                </form>
+            </div>
+        </div>
+    </div>
+    
+    <div class="col-lg-4">
+        <div class="card shadow mb-4">
+            <div class="card-header py-3">
+                <h6 class="m-0 font-weight-bold text-primary">Yardım</h6>
+            </div>
+            <div class="card-body">
+                <p>Ürün eklerken dikkat edilmesi gerekenler:</p>
+                <ul>
+                     <li>Ürün adı kısa ve anlaşılır olmalı.</li>
+                     <li>Görseller <strong>img</strong> klasöründe olmalı.</li>
+                     <li>Fiyat ondalık kısmı nokta (.) ile ayrılmalı.</li>
+                </ul>
+            </div>
+        </div>
+    </div>
+</div>
+
 @section Scripts {
-
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation\dist\jquery.validate.min.js"></script>
-
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation-unobtrusive\dist\jquery.validate.unobtrusive.min.js"></script>
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\node_modules\ckeditor\ckeditor.js"></script>
+    <script src="https://cdn.ckeditor.com/4.22.1/standard/ckeditor.js"></script>
     <script>
         CKEDITOR.replace("editor");
     </script>
-     
 }
-<h1>Yeni ürün Ekle</h1>
-<hr />
-
-<div class="row">
-    <div class="col-md-8">
-
-        @if (ViewBag.ErrorMessage != null)
-        {
-            <div class="text-danger">@ViewBag.ErrorMessage</div>
-        }
-
-
-        <form asp-controller="Admin" asp-action="CreateProduct" method="post">
-            <div asp-validation-summary="All" class="text-danger"></div>
-            <div class="form-group row">
-                <label asp-for="Name" class="col-md-2 col-form-label"> isim</label>
-                <div class="col-md-10">
-                    <input asp-for="Name" value="@Model.Name" class="form-control" />
-                    <span asp-validation-for="Name" class="text-danger"></span>
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="ImageUrl" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="ImageUrl" value="@Model.ImageUrl" class="form-control" />
-                    <span asp-validation-for="ImageUrl" class="text-danger"></span>
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="Description" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <textarea id="editor" asp-for="Description" class="form-control">
-                        value="@Model.Description"
-                    </textarea>
-                    <span asp-validation-for="Description" class="text-danger"></span>
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="Price" class="col-md-2 col-form-label">fiyat</label>
-                <div class="col-md-10">
-                    <input asp-for="Price" value="@Model.Price" class="form-control" />
-                    <span asp-validation-for="Price" class="text-danger"></span>
-                </div>
-            </div>
-            <div class="form-group row">
-                <div class="col-md-10 offset-md-2">
-                    <button type="submit" class="btn btn-primary">Ürünü Kaydet</button>
-                </div>
-            </div>
-
-        </form>
-
-    </div>
-</div>
-
-<!--
-<h1>Create Product</h1>
-<hr />
-<div class="row">
-    <div class="cold-md-8">
-        <form asp-controller="Admin" asp-action="CreateProduct" method="post">
-            <div class="form-group row">
-                <label asp-for="Name" class="col-md-2 col-form-label">Urun Ismi</label>
-                <div class="col-md-10">
-                    <input asp-for="Name" value="" class="form-control" />
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="ImageUrl" class="col-md-2 col-form-label">Image Url</label>
-                <div class="col-md-10">
-                    <input asp-for="ImageUrl"  value="" class="form-control" />
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="Description" class="col-md-2 col-form-label">Açıklama</label>
-                <div class="col-md-10">
-                    <textarea id="editor" asp-for="Description" class="form-control"></textarea>
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="Price" class="col-md-2 col-form-label">Fiyat</label>
-                <div class="col-md-10">
-                    <input asp-for="Price" value="" class="form-control" />
-                </div>
-            </div>
-            <div class="form-group row">
-
-                <div class="col-md-10 offset-md-2">
-                    <button type="submit"  class="btn btn-primary ">Ürünü Kaydet</button>
-                </div>
-            </div>
-
-        </form>
-    </div>
-</div>
-    -->
