@@ -1,75 +1,95 @@
 ﻿@model CategoryModel
 @{
-    ViewData["Title"] = "EditCategory";
+    ViewData["Title"] = "Kategori Düzenle";
 }
 
-<h1>Edit Category</h1>
-<hr />
-
-<div class="row">
-    <div class="col-md-4">
-        <form asp-controller="Admin" asp-action="EditCategory" method="post">
-            <input type="hidden" name="categoryId" value="@Model.Id" />
-            <div class="form-group row">
-                <label asp-for="Name" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Name" value="@Model.Name" class="form-control" />
-                </div>
-            </div>
-
-            <div class="form-group row">
-                <div class="col-md-10 offset-md-2">
-                    <button type="submit" class="btn btn-primary">Save Category</button>
-                </div>
-            </div>
-
-        </form>
-    </div>
-    <div class="col-md-8">
-        @if (Model.Products.Count() > 0)
-        {
-            <table class="table table-bordered">
-                <thead>
-                    <tr>
-                        <td style="width:30px;">Id</td>
-                        <td style="width:100px;">Image</td>
-                        <td>Name</td>
-                        <td style="width:12px;">Price</td>
-                        <td style="width:150px;"></td>
-                    </tr>
-                </thead>
-                <tbody>
-                    @foreach (var item in Model.Products)
-                    {
-                        <tr>
-                            <td>@item.Id</td>
-                            <td>
-                                <img src="~/img/@item.ImageUrl" width="80" />
-                            </td>
-                            <td>@item.Name</td>
-                            <td>@item.Price</td>
-                            <td>
-                                <a class="btn btn-primary btn-sm mr-2" href="/admin/products/@item.Id">Edit</a>
-                                <form action="/admin/deletefromcategory" method="post" style="display:inline;">
-                                    <input type="hidden" name="productId" value="@item.Id" />
-                                    <input type="hidden" name="categoryId" value="@Model.Id" />
-                                    <button type="submit" class="btn btn-danger btn-sm">Delete</button>
-                                </form>
-                            </td>
-                        </tr>
-                    }
-
-                </tbody>
-
-            </table>
-        }
-        else
-        {
-            <div class="alert alert-warning">
-                <h4>Bu kategoride  ürün yok</h4>
-            </div>
-        }
-    </div>
+<div class="d-flex justify-content-between align-items-center mb-4">
+    <h1 class="h3 mb-0 text-gray-800">Kategori Düzenle</h1>
+    <a href="/admin/categories" class="btn btn-outline-secondary btn-sm">
+        <i class="bi bi-arrow-left"></i> Geri Dön
+    </a>
 </div>
 
+<div class="row">
+    <div class="col-lg-4">
+        <div class="card shadow mb-4">
+            <div class="card-header py-3">
+                <h6 class="m-0 font-weight-bold text-primary">Kategori Bilgileri</h6>
+            </div>
+            <div class="card-body">
+                <form asp-controller="Admin" asp-action="EditCategory" method="post">
+                    <input type="hidden" name="categoryId" value="@Model.Id" />
+                    
+                    <div class="mb-3">
+                        <label asp-for="Name" class="form-label fw-bold">Kategori Adı</label>
+                        <input asp-for="Name" class="form-control" />
+                         <span asp-validation-for="Name" class="text-danger small"></span>
+                    </div>
 
+                    <div class="d-grid">
+                        <button type="submit" class="btn btn-primary">
+                            <i class="bi bi-save me-2"></i> Değişiklikleri Kaydet
+                        </button>
+                    </div>
+                </form>
+            </div>
+        </div>
+    </div>
+    
+    <div class="col-lg-8">
+        <div class="card shadow mb-4">
+            <div class="card-header py-3">
+                <h6 class="m-0 font-weight-bold text-primary">Bu Kategorideki Ürünler</h6>
+            </div>
+            <div class="card-body">
+                @if (Model.Products.Count() > 0)
+                {
+                    <div class="table-responsive">
+                        <table class="table table-hover align-middle">
+                            <thead class="table-light">
+                                <tr>
+                                    <th scope="col" style="width:50px;">#ID</th>
+                                    <th scope="col" style="width:80px;">Görsel</th>
+                                    <th scope="col">Ürün Adı</th>
+                                    <th scope="col" style="width:120px;">Fiyat</th>
+                                    <th scope="col" style="width:160px;">İşlemler</th>
+                                </tr>
+                            </thead>
+                            <tbody>
+                                @foreach (var item in Model.Products)
+                                {
+                                    <tr>
+                                        <td>@item.Id</td>
+                                        <td>
+                                            <img src="~/img/@item.ImageUrl" class="img-thumbnail" style="width: 50px; height: 50px; object-fit: cover;" />
+                                        </td>
+                                        <td class="fw-bold">@item.Name</td>
+                                        <td>@item.Price?.ToString("c")</td>
+                                        <td>
+                                            <a class="btn btn-primary btn-sm me-1 shadow-sm" href="/admin/products/@item.Id" title="Düzenle">
+                                                <i class="bi bi-pencil-square"></i>
+                                            </a>
+                                            <form action="/admin/deletefromcategory" method="post" class="d-inline" onsubmit="return confirm('Bu ürünü kategoriden çıkarmak istediğinize emin misiniz?');">
+                                                <input type="hidden" name="productId" value="@item.Id" />
+                                                <input type="hidden" name="categoryId" value="@Model.Id" />
+                                                <button type="submit" class="btn btn-danger btn-sm shadow-sm" title="Kategoriden Çıkar">
+                                                    <i class="bi bi-trash"></i>
+                                                </button>
+                                            </form>
+                                        </td>
+                                    </tr>
+                                }
+                            </tbody>
+                        </table>
+                    </div>
+                }
+                else
+                {
+                    <div class="alert alert-warning m-3" role="alert">
+                        <i class="bi bi-exclamation-circle me-2"></i> Bu kategoriye ait ürün bulunmamaktadır.
+                    </div>
+                }
+            </div>
+        </div>
+    </div>
+</div>
