@@ -1,96 +1,116 @@
 ﻿@model ProductModel
 @{
-    ViewData["Title"] = "EditProduct";
+    ViewData["Title"] = "Ürün Düzenle";
     var counter = 0;
 }
 
+<div class="d-flex justify-content-between align-items-center mb-4">
+    <h1 class="h3 mb-0 text-gray-800">Ürünü Düzenle</h1>
+    <a href="/admin/products" class="btn btn-outline-secondary btn-sm">
+        <i class="bi bi-arrow-left"></i> Geri Dön
+    </a>
+</div>
+
+<form asp-controller="Admin" asp-action="Edit" method="post" enctype="multipart/form-data">
+    <div asp-validation-summary="All" class="text-danger mb-3"></div>
+    <div class="row">
+        <div class="col-lg-8">
+            <div class="card shadow mb-4">
+                <div class="card-header py-3">
+                     <h6 class="m-0 font-weight-bold text-primary">Ürün Bilgileri</h6>
+                </div>
+                <div class="card-body">
+                    <input type="hidden" name="Id" value="@Model.Id" />
+                    
+                    <div class="mb-3 row">
+                        <label asp-for="Name" class="col-sm-3 col-form-label text-end fw-bold">Ürün Adı</label>
+                        <div class="col-sm-9">
+                            <input asp-for="Name" class="form-control" />
+                             <span asp-validation-for="Name" class="text-danger small"></span>
+                        </div>
+                    </div>
+
+                    <div class="mb-3 row">
+                        <label asp-for="ImageUrl" class="col-sm-3 col-form-label text-end fw-bold">Görsel</label>
+                        <div class="col-sm-9">
+                            <input type="hidden" name="ImageUrl" value="@Model.ImageUrl" />
+                            <div class="mb-2">
+                                <img src="~/img/@Model.ImageUrl" class="img-thumbnail" style="width: 120px; object-fit: cover;" />
+                            </div>
+                            <input type="file" name="file" class="form-control" />
+                        </div>
+                    </div>
+
+                    <div class="mb-3 row">
+                        <label asp-for="Price" class="col-sm-3 col-form-label text-end fw-bold">Fiyat</label>
+                        <div class="col-sm-9">
+                             <div class="input-group">
+                                <input asp-for="Price" class="form-control" type="number" step="0.01" />
+                                <span class="input-group-text">₺</span>
+                            </div>
+                            <span asp-validation-for="Price" class="text-danger small"></span>
+                        </div>
+                    </div>
+                    
+                    <div class="mb-3 row">
+                        <label asp-for="Description" class="col-sm-3 col-form-label text-end fw-bold">Açıklama</label>
+                        <div class="col-sm-9">
+                            <textarea id="editor" asp-for="Description" class="form-control">@Model.Description</textarea>
+                             <span asp-validation-for="Description" class="text-danger small"></span>
+                        </div>
+                    </div>
+
+                    <div class="row">
+                        <div class="col-sm-9 offset-sm-3">
+                            <button type="submit" class="btn btn-primary">
+                                <i class="bi bi-save me-2"></i> Değişiklikleri Kaydet
+                            </button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+        
+        <div class="col-lg-4">
+            <div class="card shadow mb-4">
+                <div class="card-header py-3">
+                    <h6 class="m-0 font-weight-bold text-primary">Kategoriler</h6>
+                </div>
+                <div class="card-body">
+                    <div class="alert alert-info py-2 small">
+                        <i class="bi bi-info-circle me-1"></i> Ürünün ait olduğu kategorileri seçiniz.
+                    </div>
+                    
+                    @foreach (var item in (List<Category>)ViewBag.Categories)
+                    {
+                        var isChecked = "";
+                        if (Model.SelectedCategories.Any(i => i.Id == item.Id))
+                        {
+                            isChecked = "checked";
+                            <input type="hidden" name="SelectedCategories[@counter].Id" value="@item.Id" />
+                            <input type="hidden" name="SelectedCategories[@counter].Name" value="@item.Name" />
+                            counter++;
+                        }
+                        
+                        <div class="form-check">
+                            <input type="checkbox"
+                                   name="categoryIds"
+                                   value="@item.Id"
+                                   class="form-check-input"
+                                   id="category@(item.Id)"
+                                   @isChecked>
+                            <label class="form-check-label" for="category@(item.Id)">@item.Name</label>
+                        </div>
+                    }
+                </div>
+            </div>
+        </div>
+    </div>
+</form>
+
 @section Scripts {
-
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation\dist\jquery.validate.min.js"></script>
-
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\modules\jquery-validation-unobtrusive\dist\jquery.validate.unobtrusive.min.js"></script>
-    <script src="C:\Users\Asil\Desktop\ShopApp\ShopApp.WebUI\node_modules\ckeditor\ckeditor.js"></script>
+    <script src="https://cdn.ckeditor.com/4.22.1/standard/ckeditor.js"></script>
     <script>
         CKEDITOR.replace("editor");
     </script>
-
 }
-<h1>Edit Product</h1>
-<hr />
-
-<form asp-controller="Admin" asp-action="Edit" method="post" enctype="multipart/form-data">
-    <div asp-validation-summary="All" class="text-danger"></div>
-    <div class="row">
-        <div class="col-md-8">
-
-            <input type="hidden" name="Id" value="@Model.Id" />
-            <div class="form-group row">
-                <label asp-for="Name" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Name" value="@Model.Name" class="form-control" />
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="ImageUrl" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input type="hidden" name="ImageUrl" value="@Model.ImageUrl" />
-                    <img src="~/img/@Model.ImageUrl" width="80" />
-                    <input type="file" name="file" value="" />
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="Description" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <textarea id="editor" asp-for="Description" class="form-control">
-                        @Model.Description
-                    </textarea>
-                </div>
-            </div>
-            <div class="form-group row">
-                <label asp-for="Price" class="col-md-2 col-form-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Price" value="@Model.Price" class="form-control" />
-                </div>
-            </div>
-            <div class="form-group row">
-                <div class="col-md-10 offset-md-2">
-                    <button type="submit" class="btn btn-primary">Save Product</button>
-                </div>
-            </div>
-
-        </div>
-        <div class="col-md-4">
-            @foreach (var item in (List<Category>)ViewBag.Categories)
-            {
-                var isChecked = "";
-                if (Model.SelectedCategories.Any(i => i.Id == item.Id))
-                {
-                    isChecked = "checked";
-
-                    <input type="hidden" name="SelectedCategories[@counter].Id" value="@item.Id" />
-                    <input type="hidden" name="SelectedCategories[@counter].Name" value="@item.Name" />
-
-                    counter++;
-
-                }
-                else
-                {
-                    isChecked = "";
-                }
-
-                <div class="custom-control custom-checkbox">
-                    <input type="checkbox"
-                           name="categoryIds"
-                           value="@item.Id"
-                           class="custom-control-input"
-                           id="category@(item.Id)"
-                           @isChecked>
-                    <label class="custom-control-label" for="category@(item.Id)">@item.Name</label>
-                </div>
-            }
-        </div>
-    </div>
-
-</form>
-
-
