@@ -1,53 +1,65 @@
 ﻿@model ProductListModel
 @{
-    ViewData["Title"] = "Index";
+    ViewData["Title"] = "Ürün Listesi";
 }
 
-<h1>Admin Ürünleri</h1>
-<hr />
+<div class="d-flex justify-content-between align-items-center mb-4">
+    <h1 class="h3 mb-0 text-gray-800">Ürün Yönetimi</h1>
+    <a href="/admin/products/create" class="btn btn-primary shadow-sm">
+        <i class="bi bi-plus-lg text-white-50"></i> Yeni Ürün Ekle
+    </a>
+</div>
 
-@if (Model.Products.Count() > 0)
-{
-
-    <table class="table table-bordered">
-        <thead>
-            <tr>
-                <td style="width:30px;">Id</td>
-                <td style="width:100px;">Resim</td>
-                <td>İsim</td>
-                <td style="width:12px;">Fiyat</td>
-                <td style="width:150px;"></td>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var item in Model.Products)
-            {
-                <tr>
-                    <td>@item.Id</td>
-                    <td>
-                        <img src="~/img/@item.ImageUrl" width="80" />
-                    </td>
-                    <td>@item.Name</td>
-                    <td>@item.Price</td>
-                    <td>
-                        <a class="btn btn-primary btn-sm mr-2" href="/admin/products/@item.Id">Edit</a>
-                        <form action="/admin/deleteproduct" method="post" style="display:inline;">
-                            <input type="hidden" name="productId" value="@item.Id" />
-                            <button type="submit" class="btn btn-danger btn-sm">Delete</button>
-                        </form>
-                    </td>
-                </tr>
-            }
-
-        </tbody>
-
-    </table>
-}
-else
-{
-    <div class="alert alert-warning">
-        <h4>No Products</h4>
+<div class="card shadow mb-4">
+    <div class="card-header py-3">
+        <h6 class="m-0 font-weight-bold text-primary">Ürün Listesi</h6>
     </div>
-}
-
-
+    <div class="card-body">
+        @if (Model.Products.Count() > 0)
+        {
+            <div class="table-responsive">
+                <table class="table table-hover align-middle">
+                    <thead class="table-light">
+                        <tr>
+                            <th scope="col" style="width:50px;">#ID</th>
+                            <th scope="col" style="width:100px;">Görsel</th>
+                            <th scope="col">Ürün Adı</th>
+                            <th scope="col" style="width:150px;">Fiyat</th>
+                            <th scope="col" style="width:180px;">İşlemler</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @foreach (var item in Model.Products)
+                        {
+                            <tr>
+                                <td>@item.Id</td>
+                                <td>
+                                    <img src="~/img/@item.ImageUrl" class="img-thumbnail" style="width: 60px; height: 60px; object-fit: cover;" />
+                                </td>
+                                <td class="fw-bold">@item.Name</td>
+                                <td>@item.Price?.ToString("c")</td>
+                                <td>
+                                    <a class="btn btn-primary btn-sm me-1 shadow-sm" href="/admin/products/@item.Id" title="Düzenle">
+                                        <i class="bi bi-pencil-square"></i>
+                                    </a>
+                                    <form action="/admin/deleteproduct" method="post" class="d-inline" onsubmit="return confirm('Bu ürünü silmek istediğinize emin misiniz?');">
+                                        <input type="hidden" name="productId" value="@item.Id" />
+                                        <button type="submit" class="btn btn-danger btn-sm shadow-sm" title="Sil">
+                                            <i class="bi bi-trash"></i>
+                                        </button>
+                                    </form>
+                                </td>
+                            </tr>
+                        }
+                    </tbody>
+                </table>
+            </div>
+        }
+        else
+        {
+            <div class="alert alert-warning m-3" role="alert">
+                <i class="bi bi-exclamation-circle me-2"></i> Sistemde kayıtlı ürün bulunamadı.
+            </div>
+        }
+    </div>
+</div>
