@@ -1,111 +1,120 @@
 ﻿@model OrderModel
 @{
-    ViewData["Title"] = "Checkout";
+    ViewData["Title"] = "Ödeme";
 }
 
-<h1>Checkout</h1>
+<h1 class="mb-4">Ödeme İşlemleri</h1>
 <hr />
 <div class="row">
     <div class="col-md-8">
-        <h4 class="mb-3">Order Details</h4>
+        <h4 class="mb-3">Kargo & Fatura Bilgileri</h4>
         <form action="/checkout" method="post">
-            <div class="row">
-                <div class="col-md-6 mb-3">
-                    <label asp-for="@Model.FirstName"></label>
-                    <input asp-for="@Model.FirstName" value="" class="form-control" />
+            <div class="row g-3">
+                <div class="col-sm-6">
+                    <label asp-for="@Model.FirstName" class="form-label">Ad</label>
+                    <input asp-for="@Model.FirstName" class="form-control" placeholder="" required />
                 </div>
-                <div class="col-md-6 mb-3">
-                    <label asp-for="@Model.LastName"></label>
-                    <input asp-for="@Model.LastName" value="" class="form-control" />
+                <div class="col-sm-6">
+                    <label asp-for="@Model.LastName" class="form-label">Soyad</label>
+                    <input asp-for="@Model.LastName" class="form-control" placeholder="" required />
                 </div>
-            </div>
-            <div class="mb-3">
-                <label asp-for="@Model.Address"></label>
-                <textarea asp-for="@Model.Address" class="form-control"></textarea>
-            </div>
-            <div class="row">
-                <div class="col-md-4 mb-3">
-                    <label asp-for="@Model.City"></label>
-                    <input asp-for="@Model.City" value="" class="form-control" />
+
+                <div class="col-12">
+                     <label asp-for="@Model.Email" class="form-label">Email</label>
+                     <input asp-for="@Model.Email" class="form-control" placeholder="siz@ornek.com" />
                 </div>
-                <div class="col-md-4 mb-3">
-                    <label asp-for="@Model.Phone"></label>
-                    <input asp-for="@Model.Phone" value="" class="form-control" />
+                 <div class="col-12">
+                     <label asp-for="@Model.Phone" class="form-label">Telefon</label>
+                     <input asp-for="@Model.Phone" class="form-control" placeholder="0555 555 55 55" />
                 </div>
-                <div class="col-md-4 mb-3">
-                    <label asp-for="@Model.Email"></label>
-                    <input asp-for="@Model.Email" value="" class="form-control" />
+
+                <div class="col-12">
+                    <label asp-for="@Model.Address" class="form-label">Adres</label>
+                    <textarea asp-for="@Model.Address" class="form-control" rows="3" placeholder="1234 Cadde Sokak" required></textarea>
                 </div>
-            </div>
-            <hr class="mb-4" />
-            <div class="custom-control custom-checkbox">
-                <input type="checkbox" class="custom-control-input" id="save-info">
-                <label class="custom-control-label" for="save-info">Save this information for next time</label>
-            </div>
-            <hr class="mb-4" />
-            <div class="d-block my-3">
-                <div class="custom-control custom-radio">
-                    <input id="credit" name="paymentMethod" type="radio" class="custom-control-input" checked="" required="">
-                    <label class="custom-control-label" for="credit">Credit card</label>
-                </div>
-                <div class="custom-control custom-radio">
-                    <input id="paypal" name="paymentMethod" type="radio" class="custom-control-input" required="">
-                    <label class="custom-control-label" for="eft">EFT</label>
+                
+                 <div class="col-md-6">
+                    <label asp-for="@Model.City" class="form-label">Şehir</label>
+                    <input asp-for="@Model.City" class="form-control" required />
                 </div>
             </div>
-            <h4 class="mb-3">Payment</h4>
-            <div class="row">
-                <div class="col-md-6 mb-3">
-                    <label asp-for="@Model.CardName"></label>
-                    <input asp-for="@Model.CardName" value="" class="form-control" />
+
+            <hr class="my-4" />
+
+            <div class="form-check">
+                <input type="checkbox" class="form-check-input" id="save-info">
+                <label class="form-check-label" for="save-info">Bilgilerimi kaydet</label>
+            </div>
+
+            <hr class="my-4" />
+
+            <h4 class="mb-3">Ödeme Yöntemi</h4>
+
+            <div class="my-3">
+                <div class="form-check">
+                    <input id="credit" name="paymentMethod" type="radio" class="form-check-input" checked required>
+                    <label class="form-check-label" for="credit">Kredi Kartı</label>
                 </div>
-                <div class="col-md-6 mb-3">
-                    <label asp-for="@Model.CardNumber"></label>
-                    <input asp-for="@Model.CardNumber" value="" class="form-control" />
+                <div class="form-check">
+                    <input id="eft" name="paymentMethod" type="radio" class="form-check-input" required>
+                    <label class="form-check-label" for="eft">Havale / EFT</label>
                 </div>
             </div>
-            <div class="row">
-                <div class="col-md-4 mb-3">
-                    <label asp-for="@Model.ExpirationMonth"></label>
-                    <input asp-for="@Model.ExpirationMonth" value="" class="form-control" />
+
+            <div class="row g-3">
+                <div class="col-md-6">
+                    <label asp-for="@Model.CardName" class="form-label">Kart Üzerindeki İsim</label>
+                    <input asp-for="@Model.CardName" class="form-control" required />
+                    <small class="text-muted">Kartın ön yüzündeki isim</small>
                 </div>
-                <div class="col-md-4 mb-3">
-                    <label asp-for="@Model.ExpirationYear"></label>
-                    <input asp-for="@Model.ExpirationYear" value="" class="form-control" />
+                <div class="col-md-6">
+                    <label asp-for="@Model.CardNumber" class="form-label">Kart Numarası</label>
+                    <input asp-for="@Model.CardNumber" class="form-control" required />
                 </div>
-                <div class="col-md-4 mb-3">
-                    <label asp-for="@Model.Cvv"></label>
-                    <input asp-for="@Model.Cvv" value="" class="form-control" />
+                <div class="col-md-3">
+                    <label asp-for="@Model.ExpirationMonth" class="form-label">Ay</label>
+                    <input asp-for="@Model.ExpirationMonth" class="form-control" placeholder="MM" required />
+                </div>
+                 <div class="col-md-3">
+                    <label asp-for="@Model.ExpirationYear" class="form-label">Yıl</label>
+                    <input asp-for="@Model.ExpirationYear" class="form-control" placeholder="YY" required />
+                </div>
+                <div class="col-md-3">
+                    <label asp-for="@Model.Cvv" class="form-label">CVV</label>
+                    <input asp-for="@Model.Cvv" class="form-control" required />
                 </div>
             </div>
-            <hr class="mb-4" />
-            <button type="submit" class="btn btn-primary btn-lg btn-block">Complete Order</button>
+
+            <hr class="my-4" />
+
+            <button type="submit" class="w-100 btn btn-primary btn-lg">siparişi Tamamla</button>
         </form>
     </div>
-    <div class="col-md-4 mb-4">
-
-        <h4 class="mb-3">
-            <span class="text-muted">Your Cart</span>
-        </h4>
-        <ul class="list-group mb-3">
-            @foreach (var item in Model.CartModel.CartItems)
-            {
-                <li class="list-group-item d-flex justify-content-between">
-                    <div>
-                        <h6 class="my-0">
-                            @item.Name
-                        <small>x @item.Quantity</small>
-                    </h6>
-                </div>
-                <span class="text-muted">@item.Price TL</span>
-            </li>
-        }
-            <li class="list-group-item d-flex justify-content-between">
-                <span>Total (TL)</span>
-                <strong>@Model.CartModel.TotalPrice() TL</strong>
-            </li>
-        </ul>
-
+    <div class="col-md-4">
+         @if (Model.CartModel?.CartItems?.Any() == true)
+         {
+            <div class="position-sticky" style="top: 6rem;">
+                <h4 class="d-flex justify-content-between align-items-center mb-3">
+                    <span class="text-primary">Sepetiniz</span>
+                    <span class="badge bg-primary rounded-pill">@Model.CartModel.CartItems.Count</span>
+                </h4>
+                <ul class="list-group mb-3 shadow-sm">
+                    @foreach (var item in Model.CartModel.CartItems)
+                    {
+                        <li class="list-group-item d-flex justify-content-between lh-sm">
+                            <div>
+                                <h6 class="my-0">@item.Name</h6>
+                                <small class="text-muted">Adet: @item.Quantity</small>
+                            </div>
+                            <span class="text-muted">@item.Price.ToString("c")</span>
+                        </li>
+                    }
+                    <li class="list-group-item d-flex justify-content-between fw-bold bg-light">
+                        <span>Toplam (TL)</span>
+                        <strong>@Model.CartModel.TotalPrice().ToString("c")</strong>
+                    </li>
+                </ul>
+            </div>
+         }
     </div>
 </div>
-
