@@ -1,27 +1,36 @@
 ﻿@model Product
 
-<div class="col-md-4">
-    <div class="card mb-3">
-        <img src="~/img/@Model.ImageUrl" alt="@Model.Name" />
-        <div class="card-body">
-            <h4 class="card-title">
-                <a href="#">@Model.Name</a>
-            </h4>
-            <small class="text-muted">
-                <i class="fas fa-lira-sign"></i> @Model.Price
-            </small>
+<div class="col">
+    <div class="card h-100 shadow-sm product-card border-0">
+        <div class="position-relative overflow-hidden p-3 text-center">
+            <img src="~/img/@Model.ImageUrl" class="card-img-top img-fluid" alt="@Model.Name" style="height: 200px; object-fit: contain;" />
         </div>
-        <div class="card-footer text-center">
-            <a asp-controller="Shop" asp-action="Details" asp-route-id="@Model.Id" class="btn btn-outline-primary btn-sm">Detaylar</a>
-            @if (User.Identity.IsAuthenticated)
+        <div class="card-body d-flex flex-column">
+            <h5 class="card-title text-truncate" title="@Model.Name">
+                <a asp-controller="Shop" asp-action="Details" asp-route-id="@Model.Id" class="text-decoration-none text-dark fw-bold stretched-link">
+                    @Model.Name
+                </a>
+            </h5>
+            <div class="mt-auto d-flex justify-content-between align-items-center">
+                <span class="fs-5 fw-bold text-success">
+                    @(Model.Price?.ToString("N2")) <i class="bi bi-currency-lira"></i>
+                </span>
+            </div>
+        </div>
+        <div class="card-footer bg-white border-top-0 d-flex justify-content-between align-items-center pb-3">
+             <a asp-controller="Shop" asp-action="Details" asp-route-id="@Model.Id" class="btn btn-outline-secondary btn-sm position-relative z-2">
+                <i class="bi bi-eye"></i> İncele
+            </a>
+            @if (User.Identity?.IsAuthenticated == true)
             {
-                <form asp-controller="Cart" asp-action="AddToCart" method="post" style="display:inline;">
+                <form asp-controller="Cart" asp-action="AddToCart" method="post" class="d-inline position-relative z-2">
                     <input type="hidden" name="productId" value="@Model.Id" />
                     <input type="hidden" name="quantity" value="1" />
-                    <button type="submit" class="btn btn-outline-primary btn-sm ml-3">Sepete Ekle</button>
+                    <button type="submit" class="btn btn-primary btn-sm btn-add-cart shadow-sm">
+                        <i class="bi bi-cart-plus"></i> Ekle
+                    </button>
                 </form>
             }
-
         </div>
     </div>
 </div>
