@@ -1,30 +1,30 @@
-﻿<div id="mySlider" class="carousel slide mb-3" data-ride="carousel">
+﻿<div id="mySlider" class="carousel slide mb-3 shadow rounded overflow-hidden" data-bs-ride="carousel">
 
-    <ol class="carousel-indicators">
-        <li data-target="#mySlider" data-slide-to="0" class="active"></li>
-        <li data-target="#mySlider" data-slide-to="1"></li>
-        <li data-target="#mySlider" data-slide-to="2"></li>
-    </ol>
+    <div class="carousel-indicators">
+        <button type="button" data-bs-target="#mySlider" data-bs-slide-to="0" class="active" aria-current="true" aria-label="Slide 1"></button>
+        <button type="button" data-bs-target="#mySlider" data-bs-slide-to="1" aria-label="Slide 2"></button>
+        <button type="button" data-bs-target="#mySlider" data-bs-slide-to="2" aria-label="Slide 3"></button>
+    </div>
 
     <div class="carousel-inner">
         <div class="carousel-item active">
-            <img src="~/img/slider1.png" class="d-block img-fluid" />
+            <img src="~/img/slider1.png" class="d-block w-100" alt="Slider 1" style="height: 300px; object-fit: cover;" />
         </div>
         <div class="carousel-item">
-            <img src="~/img/slider2.png" class="d-block img-fluid" />
+            <img src="~/img/slider2.png" class="d-block w-100" alt="Slider 2" style="height: 300px; object-fit: cover;" />
         </div>
         <div class="carousel-item">
-            <img src="~/img/slider3.png" class="d-block img-fluid" />
+            <img src="~/img/slider3.png" class="d-block w-100" alt="Slider 3" style="height: 300px; object-fit: cover;" />
         </div>
     </div>
 
-    <a href="#mySlider" data-slide="prev" class="carousel-control-prev">
-        <span class="carousel-control-prev-icon"></span>
-    </a>
-
-    <a href="#mySlider" data-slide="next" class="carousel-control-next">
-        <span class="carousel-control-next-icon"></span>
-    </a>
-
+    <button class="carousel-control-prev" type="button" data-bs-target="#mySlider" data-bs-slide="prev">
+        <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+        <span class="visually-hidden">Previous</span>
+    </button>
+    <button class="carousel-control-next" type="button" data-bs-target="#mySlider" data-bs-slide="next">
+        <span class="carousel-control-next-icon" aria-hidden="true"></span>
+        <span class="visually-hidden">Next</span>
+    </button>
 
 </div>
