@@ -1,22 +1,19 @@
 ﻿@model ProductListModel
+
 <div class="row">
-    <div class="col-md-3">
+    <div class="col-lg-3 mb-4">
         @await Component.InvokeAsync("CategoryList")
     </div>
-    <div class="col-md-9">
-
-       
-
-        <div class="row">
+    <div class="col-lg-9">
+        <div class="row row-cols-1 row-cols-md-2 row-cols-lg-3 g-4">
             @foreach (var product in Model.Products)
             {
                 <partial name="_product" for="@product" />
             }
         </div>
-        <div page-model="@Model.PageInfo">
-
+        
+        <div class="mt-4 d-flex justify-content-center">
+            <div page-model="@Model.PageInfo"></div>
         </div>
-
     </div>
 </div>
-
